discounting: Discounting of Affect

Functions created to aid with a large-scale evaluation of discounting models for capturing affect in experimental and daily-life datasets. Introduces a class for each separate model as well as a class for datasets, making every function internal to the package run within a standardized framework

See also