Fit a model to data

This method enables the user to fit a model (defined in model) on a particular dataset (defined in data). The estimation procedure makes use of numerical optimization using either DEoptim or nloptr, as specified by the user. Estimation proceeds through an optimization according to the output of the objective function of the provided model, as defined through objective_function, thus using least-squares as optimization standard.

Usage

fit(model, data, ...)

# S4 method for class 'model,dataset'
fit(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "optim",
  lower = NULL,
  upper = NULL,
  method = "Nelder-Mead",
  na.rm = TRUE,
  ...
)

Arguments

model

Instance of the model-class, defining the model to evaluate the objective function for.

data

Instance of the dataset-class containing the data to fit the model to.

...

Arguments passed on to the control parameters of the optimizer, either to DEoptim.control, the opts argument of nloptr, the control list of the optim function, or to the additional arguments of your own optimizer function (see Details).

dynamics

Character denoting the structure of the dynamical matrices. Can either be "anisotropic" (completely free), "symmetric" (symmetric around the diagonal), and "isotropic" (diagonal). Note that this influences different parameters for different models, namely \(\Gamma\) for the exponential discounting model, \(N\) and \(K\) for the quasi-hyperbolic discounting model, and \(\Gamma\) and \(N\) for the double-exponential discounting model. Defaults to "isotropic".

covariance

Character denoting the structure of covariance matrix. Can either by "symmetric" (symmetric around the diagonal) and "isotropic" (diagonal). Defaults to "symmetric".

optimizer

Character denoting the optimizer to use for the estimation. Can either be "DEoptim" for the differential evolution algorithm in DEoptim, "nloptr" for the library implemented in nloptr, or "optim" for the default optimization library in R. Defaults to "optim". Alternatively, optimizer can also be a custom function that defines the estimation routine. To see what this function should look like, we refer the interested reader to the details.

lower, upper

Numeric vector containing lower and upper bounds for the parameters in the estimation routine. Uses the same defaults as get_bounds.

method

Character denoting the method to be used for estimation. Ignored if optimizer is not "optim", in which case the native arguments of the other optimization packages can be used.

na.rm

Logical denoting whether to remove NAs in the calculation of the sum squared error. Defaults to TRUE.

Value

An named list containing an instance of the model-class with the estimated parameters ("model"), the results of the optimization procedure ("fit"), the value of the objective after ending the optimization procedure ("objective"), the residuals of the model ("residuals"), and a named vector containing the values of the estimated parameters linked to a character vector explaining their content ("parameters").

Details

Note that currently, least-squares estimation is assumed, meaning that the covariance matrix is left out of the objective function. If maximum-likelihood were needed instead, then this function would need to change.

By default, the fit function performs optimizations through the optim function in R, alternatively allowing users to make use of the DEoptim or nloptr packages instead. For more difficult models, however, it may be useful to have greater control over the optimization procedure, for example by specifying your own initial conditions, by using a combination of different optimization algorithms at once, or by using another library not provided through this package.

In these cases, one can provide a function to the argument optimizer instead of a string. This function should at least take in 3 argument: A function representing the objective function, a lower bound for the parameters, and an upper bound for these same parameters (cf. optim, DEoptim, and nloptr). This function can furthermore take in additional argument, for example specifying the maximal number of iterations or controlling other aspects of the estimation routine. These can be specified in the fit function through the optional arguments ....

The custom optimization function should, minimally, output a named list with slots "parameters" and "objective", containing the final result of the parameter estimation and the value of the objective function for this final parameter set respectively. These will then be used by the fit function as final results of the optimization procedure.

Examples

# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  optimizer = "DEoptim",
  itermax = 50,
  trace = FALSE
)
#> 
#> DEoptim package
#> Differential Evolution algorithm in R
#> Authors: D. Ardia, K. Mullen, B. Peterson and J. Ulrich
#> $model
#> Model of class "exponential":
#> 
#> Dimension: 2
#> Number of predictors: 2
#> Number of parameters: 10
#> 
#> Parameters:
#>   alpha: |  0.9005992  |
#>          |  -0.9498439  |
#> 
#>   beta: | 1.756227  1.797845 |
#>         | 1.836261  1.947178 |
#> 
#>   gamma: | 0.6002239  0.00 |
#>          | 0.00  0.5047709 |
#> 
#> 
#> Covariance: | 1.28925  0.00 |
#>             | 0.00  1.561412 |
#> 
#> $fit
#> $optim
#> $optim$bestmem
#>       par1       par2       par3       par4       par5       par6       par7 
#>  0.9005992 -0.9498439  1.7562273  1.8362609  1.7978450  1.9471778  0.6002239 
#>       par8 
#>  0.5047709 
#> 
#> $optim$bestval
#> [1] 290.6292
#> 
#> $optim$nfeval
#> [1] 4080
#> 
#> $optim$iter
#> [1] 50
#> 
#> 
#> $member
#> $member$lower
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>   -1   -1   -5   -5   -5   -5    0    0 
#> 
#> $member$upper
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>    1    1    5    5    5    5    1    1 
#> 
#> $member$bestmemit
#>          par1       par2     par3      par4     par5     par6         par7
#> 1  0.06854432 -0.8842157 2.361248 1.4000135 3.250305 1.400494 0.5915340004
#> 2  0.03226230 -0.2731148 2.657543 0.8269391 1.177591 2.080725 0.6515574147
#> 3  0.03226230 -0.2731148 2.657543 0.8269391 1.177591 2.080725 0.6515574147
#> 4  0.03226230 -0.2731148 2.657543 0.8269391 1.177591 2.080725 0.6515574147
#> 5  0.01493175 -0.3089416 2.015266 1.6928262 1.733304 2.065375 0.0005517474
#> 6  0.01493175 -0.3089416 2.015266 1.6928262 1.733304 2.065375 0.0005517474
#> 7  0.06854432 -0.8842157 2.361248 1.4000135 3.250305 1.400494 0.2393551449
#> 8  0.06854432 -0.8842157 2.361248 1.4000135 3.250305 1.400494 0.2393551449
#> 9  0.88716374 -0.8103968 1.858222 2.4604005 2.563283 3.056306 0.5254040731
#> 10 0.56009005 -0.8683236 1.545935 2.6532981 1.633402 1.925104 0.3638451880
#> 11 0.78678473 -0.9390644 1.962691 1.4956971 2.046742 2.662657 0.3698532591
#> 12 0.78678473 -0.9390644 1.962691 1.4956971 2.046742 2.662657 0.3698532591
#> 13 0.78678473 -0.9390644 1.962691 1.4956971 2.046742 2.469047 0.3698532591
#> 14 0.78678473 -0.9390644 1.962691 1.4956971 2.046742 2.469047 0.3698532591
#> 15 0.78678473 -0.9390644 1.961377 1.4956971 2.046742 2.469047 0.3698532591
#> 16 0.78678473 -0.9390644 1.961377 1.4956971 2.046742 2.469047 0.3698532591
#> 17 0.78678473 -0.9390644 1.961377 1.4956971 2.046742 2.469047 0.3698532591
#> 18 0.78678473 -0.9390644 1.961377 1.4956971 2.046742 2.469047 0.3698532591
#> 19 0.78678473 -0.9390644 1.961377 1.8395561 2.046742 2.469047 0.3698532591
#> 20 0.78678473 -0.9390644 1.961377 1.8395561 2.046742 2.469047 0.3698532591
#> 21 0.78678473 -0.9390644 1.961377 1.8395561 2.046742 2.469047 0.3698532591
#> 22 0.73021898 -0.5819510 1.925423 1.2974477 1.805838 2.235467 0.5258088984
#> 23 0.73021898 -0.5819510 1.925423 1.2974477 1.805838 2.235467 0.5258088984
#> 24 0.74664034 -0.8934961 1.113810 1.6497591 1.774533 2.235312 0.5878197562
#> 25 0.99519482 -0.7370203 1.849201 2.0990616 1.601318 2.179491 0.5500116121
#> 26 0.99519482 -0.7370203 1.849201 2.0990616 1.601318 2.179491 0.5500116121
#> 27 0.99519482 -0.7370203 1.849201 2.0990616 1.601318 2.179491 0.5500116121
#> 28 0.59370464 -0.7121474 1.778439 1.8603709 1.574244 2.003089 0.5994148907
#> 29 0.59370464 -0.7535871 1.778439 1.8603709 1.574244 2.003089 0.5994148907
#> 30 0.59370464 -0.7535871 1.778439 1.8603709 1.574244 2.003089 0.5994148907
#> 31 0.59370464 -0.7535871 1.778439 1.8603709 1.574244 2.003089 0.5994148907
#> 32 0.84250000 -0.7581388 2.091590 1.9874332 2.208839 1.898497 0.5107463680
#> 33 0.82250426 -0.8297948 1.685852 1.9331077 2.065541 2.162442 0.5128311868
#> 34 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 35 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 36 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 37 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 38 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 39 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 40 0.72309877 -0.9626593 1.640667 1.9231712 1.900004 1.778870 0.5484041527
#> 41 0.80549523 -0.9026127 1.710906 1.9778655 1.903742 1.938759 0.5714691971
#> 42 0.82070381 -0.9767206 1.933628 1.9616527 1.856505 1.932444 0.5901793117
#> 43 0.82070381 -0.9767206 1.933628 1.9616527 1.856505 1.932444 0.5901793117
#> 44 0.82070381 -0.9767206 1.933628 1.9616527 1.856505 1.932444 0.5901793117
#> 45 0.82070381 -0.9767206 1.933628 1.9616527 1.856505 1.932444 0.5901793117
#> 46 0.94148876 -0.9759001 1.764254 1.9171578 1.787263 2.100868 0.5647926179
#> 47 0.90059915 -0.9498439 1.725732 1.8362609 1.797845 1.947178 0.6002238977
#> 48 0.90059915 -0.9498439 1.725732 1.8362609 1.797845 1.947178 0.6002238977
#> 49 0.90059915 -0.9498439 1.725732 1.8362609 1.797845 1.947178 0.6002238977
#> 50 0.90059915 -0.9498439 1.756227 1.8362609 1.797845 1.947178 0.6002238977
#>         par8
#> 1  0.1325339
#> 2  0.3957481
#> 3  0.3957481
#> 4  0.3957481
#> 5  0.5690431
#> 6  0.5690431
#> 7  0.7383549
#> 8  0.7383549
#> 9  0.2535781
#> 10 0.2685956
#> 11 0.5814674
#> 12 0.5814674
#> 13 0.5814674
#> 14 0.5814674
#> 15 0.5814674
#> 16 0.5814674
#> 17 0.5814674
#> 18 0.5814674
#> 19 0.5814674
#> 20 0.5814674
#> 21 0.5814674
#> 22 0.4534670
#> 23 0.4534670
#> 24 0.4112987
#> 25 0.3393013
#> 26 0.3393013
#> 27 0.3393013
#> 28 0.5629280
#> 29 0.5629280
#> 30 0.5629280
#> 31 0.5629280
#> 32 0.5439996
#> 33 0.5642967
#> 34 0.5747164
#> 35 0.5747164
#> 36 0.5747164
#> 37 0.5747164
#> 38 0.5747164
#> 39 0.5747164
#> 40 0.5747164
#> 41 0.5306107
#> 42 0.5151218
#> 43 0.5151218
#> 44 0.5151218
#> 45 0.5151218
#> 46 0.4800508
#> 47 0.5047709
#> 48 0.5047709
#> 49 0.5047709
#> 50 0.5047709
#> 
#> $member$bestvalit
#>  [1] 936.1299 919.5943 919.5943 919.5943 762.2142 762.2142 734.4692 734.4692
#>  [9] 518.6366 516.9245 474.3000 474.3000 433.2525 433.2525 433.2261 433.2261
#> [17] 433.2261 433.2261 420.8073 420.8073 420.8073 410.8874 410.8874 386.6788
#> [25] 354.1924 354.1924 354.1924 342.3828 338.1591 338.1591 338.1591 337.7880
#> [33] 330.3963 308.4895 308.4895 308.4895 308.4895 308.4895 308.4895 308.4895
#> [41] 297.8627 295.6575 295.6575 295.6575 295.6575 294.2420 290.7640 290.7640
#> [49] 290.7640 290.6292
#> 
#> $member$pop
#>            [,1]       [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.7914428 -0.9811945 1.721988 1.870203 1.892569 1.867340 0.5786754
#>  [2,] 0.9873164 -0.9251582 1.637604 1.752780 2.035505 1.645019 0.5375831
#>  [3,] 0.9642020 -0.9863900 1.780898 1.983189 1.792276 2.173803 0.5467160
#>  [4,] 0.8557458 -0.9761643 1.873303 1.790691 1.808688 1.987592 0.6214005
#>  [5,] 0.9319100 -0.8625146 1.749688 1.425667 1.573487 1.695275 0.5610951
#>  [6,] 0.9340409 -0.9606355 1.685852 1.933108 2.065541 2.162442 0.5836940
#>  [7,] 0.8623426 -0.8938227 1.781108 1.863107 1.549405 2.119663 0.6447936
#>  [8,] 0.7133707 -0.9591360 1.882370 1.881147 1.759541 1.860934 0.5577234
#>  [9,] 0.8320667 -0.9985499 1.739726 1.955841 1.956724 1.779731 0.5754154
#> [10,] 0.8412172 -0.8728242 1.875090 1.949782 1.991131 1.909857 0.6037362
#> [11,] 0.8595194 -0.9562280 1.753664 2.024559 1.802030 2.036886 0.6066689
#> [12,] 0.9045035 -0.9463616 1.551517 1.596516 2.057193 1.793371 0.6422560
#> [13,] 0.9652148 -0.9684101 1.939782 1.984259 1.575684 2.292596 0.5985991
#> [14,] 0.9572144 -0.9829065 1.853417 1.716474 1.832872 1.802562 0.5741548
#> [15,] 0.8809425 -0.9066437 1.821744 2.036365 1.694686 1.736454 0.6036528
#> [16,] 0.9568697 -0.9983886 1.928604 1.739528 1.735604 2.091449 0.6287962
#> [17,] 0.8911684 -0.9316393 1.370319 1.933683 1.920767 2.056882 0.5980291
#> [18,] 0.8528395 -0.9601045 1.800015 1.589074 1.726360 1.923855 0.5013474
#> [19,] 0.5756078 -0.8769458 1.847849 1.971718 2.017119 2.251190 0.5636984
#> [20,] 0.6856786 -0.9298277 1.850914 1.923116 1.774533 2.235312 0.5987638
#> [21,] 0.8871251 -0.9970930 1.605561 2.240458 1.821641 1.985626 0.5753788
#> [22,] 0.8601290 -0.8272035 1.686776 2.172120 1.945089 2.078950 0.5318315
#> [23,] 0.8602349 -0.9734019 1.787403 1.996750 1.751284 1.951640 0.5814492
#> [24,] 0.8329133 -0.9552969 1.677640 1.706059 1.766749 1.686970 0.6426321
#> [25,] 0.7670787 -0.8010237 1.990840 1.981158 1.890507 1.721022 0.5483070
#> [26,] 0.9648678 -0.9843986 1.964678 1.982425 1.652285 2.133915 0.6422674
#> [27,] 0.9212044 -0.9314060 1.752736 1.805752 1.865788 2.029967 0.5781224
#> [28,] 0.8756238 -0.9002483 1.818046 1.509110 1.758186 1.938241 0.5837866
#> [29,] 0.8498454 -0.8318819 1.555962 2.112991 2.170496 1.963656 0.5435419
#> [30,] 0.9502369 -0.9613186 1.592174 1.768679 1.916100 1.972165 0.5765476
#> [31,] 0.3923333 -0.9812247 1.700381 2.230127 2.018911 2.105254 0.6336182
#> [32,] 0.8425000 -0.9072014 1.769528 1.987433 2.208839 1.948969 0.5912688
#> [33,] 0.8826790 -0.9569356 1.748627 1.974656 1.772697 1.739069 0.6309354
#> [34,] 0.9030187 -0.8657588 1.832510 1.831805 2.506079 1.735843 0.5051370
#> [35,] 0.9492132 -0.9182879 1.852022 1.507266 2.158530 2.006243 0.5287362
#> [36,] 0.8645320 -0.8153497 1.753787 2.013851 1.889345 1.850400 0.6142645
#> [37,] 0.9855567 -0.9047755 1.615845 2.066192 2.301781 2.026108 0.6052533
#> [38,] 0.9949473 -0.9789621 1.688905 1.983279 1.971334 2.069566 0.6162259
#> [39,] 0.9302133 -0.9722674 1.739980 1.941726 1.930897 1.775508 0.5668260
#> [40,] 0.8793203 -0.9099422 1.774258 1.962693 1.935031 1.850787 0.5782723
#> [41,] 0.9664590 -0.9767157 1.793087 1.968402 1.965705 1.727695 0.5527097
#> [42,] 0.7400093 -0.8229684 1.516519 1.937535 1.912102 1.860201 0.6157599
#> [43,] 0.8285148 -0.9405526 1.627813 1.825066 1.716069 1.851567 0.6020606
#> [44,] 0.9414888 -0.9759001 1.764254 1.917158 1.787263 2.100868 0.5647926
#> [45,] 0.9366668 -0.9283815 1.703883 1.762551 1.871369 1.852843 0.5941945
#> [46,] 0.9398457 -0.8479381 1.894389 1.615505 2.018938 2.294177 0.6107460
#> [47,] 0.8526526 -0.9093594 1.786836 1.725160 1.840161 2.097215 0.6003616
#> [48,] 0.9554975 -0.9923363 1.718707 1.962434 2.130347 1.901181 0.5831991
#> [49,] 0.8892618 -0.9176955 1.866442 1.901930 1.963821 1.613430 0.5688468
#> [50,] 0.9017105 -0.9957556 1.858222 1.923003 1.991854 2.017321 0.6050760
#> [51,] 0.8054952 -0.9026127 1.710906 1.977866 1.903742 1.938759 0.5714692
#> [52,] 0.7149007 -0.9626496 1.908373 1.933427 1.923222 1.973361 0.5784863
#> [53,] 0.6074820 -0.9873059 1.870510 1.791410 1.927169 1.829144 0.5872564
#> [54,] 0.9233357 -0.9272127 1.737296 1.681895 1.922603 1.659345 0.6009080
#> [55,] 0.6591116 -0.7647279 1.629340 1.761010 1.840555 1.753310 0.5705847
#> [56,] 0.8704756 -0.9516786 1.779088 1.893163 1.608048 1.773104 0.6490605
#> [57,] 0.8420352 -0.9256552 1.853377 1.748764 1.838592 2.085991 0.5628326
#> [58,] 0.8485084 -0.9833555 1.750024 2.036073 1.854998 1.492940 0.6010476
#> [59,] 0.9005992 -0.9498439 1.756227 1.836261 1.797845 1.947178 0.6002239
#> [60,] 0.8405130 -0.8380431 1.984053 2.226795 1.924905 2.132693 0.5663574
#> [61,] 0.8848341 -0.9165791 1.736680 1.904452 1.719175 1.981149 0.5803274
#> [62,] 0.9317450 -0.9493250 1.760317 1.993371 1.606661 2.040996 0.5424309
#> [63,] 0.9291518 -0.9767206 1.669991 1.961653 1.856505 1.932444 0.5841846
#> [64,] 0.8995389 -0.9843610 1.625115 2.062806 1.750270 1.904099 0.5646199
#> [65,] 0.9845890 -0.9560761 1.747880 2.194163 1.568077 1.686394 0.5943906
#> [66,] 0.8288896 -0.9864997 1.694681 1.903839 2.100993 2.055655 0.5649354
#> [67,] 0.8132683 -0.9078263 1.763254 1.739550 1.908428 1.730848 0.5235420
#> [68,] 0.8175148 -0.9609412 1.796432 2.083069 2.049608 2.201816 0.5287200
#> [69,] 0.9121070 -0.9428882 1.746949 1.923171 2.044572 1.850200 0.5671680
#> [70,] 0.7877496 -0.9390644 1.857995 1.839556 2.046742 1.888306 0.5038858
#> [71,] 0.7407603 -0.9678054 1.725397 2.137512 2.097830 2.049737 0.5477624
#> [72,] 0.7853607 -0.9857283 1.753971 1.958847 1.921929 1.949354 0.6180184
#> [73,] 0.8327371 -0.6453410 1.971055 2.138612 2.042964 2.018228 0.5903207
#> [74,] 0.7296560 -0.8963143 1.678452 1.727562 1.886357 1.771896 0.6428414
#> [75,] 0.8003801 -0.9609985 1.801344 2.069703 2.030304 1.792578 0.5674021
#> [76,] 0.7288554 -0.8909332 1.919793 1.850243 1.913904 1.981090 0.5651641
#> [77,] 0.8248272 -0.9343253 1.751839 1.750370 1.786927 1.526565 0.5765589
#> [78,] 0.9528151 -0.7994200 1.648579 1.650979 1.808527 2.075144 0.5466991
#> [79,] 0.9246401 -0.9168106 1.764601 1.860371 1.574244 2.003089 0.5994149
#> [80,] 0.8951606 -0.8289865 1.739391 2.006381 1.795019 1.711152 0.5804886
#>            [,8]
#>  [1,] 0.5001956
#>  [2,] 0.5846245
#>  [3,] 0.5270454
#>  [4,] 0.5556485
#>  [5,] 0.6329644
#>  [6,] 0.4716197
#>  [7,] 0.4637514
#>  [8,] 0.4631353
#>  [9,] 0.4076155
#> [10,] 0.4858650
#> [11,] 0.4367569
#> [12,] 0.5911828
#> [13,] 0.4610656
#> [14,] 0.5459879
#> [15,] 0.4286609
#> [16,] 0.5298969
#> [17,] 0.5580755
#> [18,] 0.4935850
#> [19,] 0.4324556
#> [20,] 0.4112987
#> [21,] 0.5142098
#> [22,] 0.5238379
#> [23,] 0.4224257
#> [24,] 0.5762190
#> [25,] 0.5249721
#> [26,] 0.4910867
#> [27,] 0.5233119
#> [28,] 0.5424903
#> [29,] 0.5500549
#> [30,] 0.4706272
#> [31,] 0.4184276
#> [32,] 0.5439996
#> [33,] 0.5164578
#> [34,] 0.5373669
#> [35,] 0.5148347
#> [36,] 0.4629511
#> [37,] 0.5936635
#> [38,] 0.5073338
#> [39,] 0.5560540
#> [40,] 0.5690431
#> [41,] 0.5175328
#> [42,] 0.4679044
#> [43,] 0.5523739
#> [44,] 0.4800508
#> [45,] 0.5585133
#> [46,] 0.5185618
#> [47,] 0.4243739
#> [48,] 0.5639304
#> [49,] 0.5476437
#> [50,] 0.4886095
#> [51,] 0.5306107
#> [52,] 0.4579510
#> [53,] 0.5345285
#> [54,] 0.5500779
#> [55,] 0.5388624
#> [56,] 0.5423954
#> [57,] 0.6384893
#> [58,] 0.4971249
#> [59,] 0.5047709
#> [60,] 0.4289434
#> [61,] 0.6082218
#> [62,] 0.4793692
#> [63,] 0.5151218
#> [64,] 0.4825794
#> [65,] 0.4556991
#> [66,] 0.4103579
#> [67,] 0.5968932
#> [68,] 0.4685176
#> [69,] 0.5356805
#> [70,] 0.5512798
#> [71,] 0.5759438
#> [72,] 0.5213331
#> [73,] 0.4439279
#> [74,] 0.4743707
#> [75,] 0.4807329
#> [76,] 0.5296719
#> [77,] 0.5851313
#> [78,] 0.5282964
#> [79,] 0.4856772
#> [80,] 0.5510680
#> 
#> $member$storepop
#> list()
#> 
#> 
#> attr(,"class")
#> [1] "DEoptim"
#> 
#> $objective
#> [1] 290.6292
#> 
#> $residuals
#>               [,1]        [,2]
#>   [1,]  1.20897279 -1.19082658
#>   [2,] -0.32712661 -0.71767159
#>   [3,]  1.69252238  1.19995309
#>   [4,] -0.77324599 -0.47352348
#>   [5,] -2.18916264 -0.49749181
#>   [6,]  1.29924819  0.73491787
#>   [7,]  0.04862331  1.06917042
#>   [8,]  0.29451321  0.63211704
#>   [9,]  2.94177748  2.42494219
#>  [10,]  0.45992818 -0.16174663
#>  [11,] -0.59212289 -1.36696871
#>  [12,]  1.52692714  1.65780395
#>  [13,] -0.68736967 -0.79620298
#>  [14,]  0.53191152 -0.73580669
#>  [15,] -0.25733163 -1.95036337
#>  [16,] -1.27224990  1.34231332
#>  [17,] -0.98526394 -2.83435262
#>  [18,] -0.71228084 -1.12544274
#>  [19,]  0.55653722  0.24925534
#>  [20,]  1.60986663  1.36907742
#>  [21,]  2.16424276  2.77637045
#>  [22,] -1.25327712 -0.52407109
#>  [23,] -1.82650660 -2.50074034
#>  [24,]  1.73706773  1.60694562
#>  [25,]  2.30045521  1.11029140
#>  [26,]  0.48741090 -2.22517254
#>  [27,] -1.44037003 -0.66983350
#>  [28,] -1.76469522 -0.57739495
#>  [29,]  0.90001697 -0.20958047
#>  [30,] -1.02876801 -2.30130111
#>  [31,]  0.63972779 -1.24075761
#>  [32,] -0.88114477 -2.22261073
#>  [33,]  0.48562270  1.25269830
#>  [34,] -0.22736201 -1.97928664
#>  [35,]  0.15814129 -1.27220299
#>  [36,]  0.65613872  1.72198585
#>  [37,] -1.10427186  0.49112285
#>  [38,]  0.69152520  0.46526471
#>  [39,]  0.66877794  1.72917305
#>  [40,]  0.80856930 -0.39520600
#>  [41,]  0.23011493  0.70623532
#>  [42,] -0.06149882  0.02432103
#>  [43,]  0.52251762 -1.29041160
#>  [44,] -0.84647049 -1.36940694
#>  [45,]  2.02168840 -0.60744846
#>  [46,]  1.00013695  0.76348957
#>  [47,] -1.15595404  0.47300156
#>  [48,]  1.69425106 -0.16587603
#>  [49,] -1.79987326  0.26389073
#>  [50,] -0.58723298 -1.16321329
#>  [51,]  1.39024313  0.99987952
#>  [52,]  0.11395395  0.50596791
#>  [53,] -0.28289506 -0.58965433
#>  [54,]  0.52343122  0.41232558
#>  [55,]  1.14650104  1.41872769
#>  [56,] -0.19980746  0.96464907
#>  [57,]  1.97554014  1.29645474
#>  [58,] -1.00738466  1.44412366
#>  [59,] -2.04255724 -0.61841037
#>  [60,] -0.66446067 -0.05478740
#>  [61,] -0.70188270 -2.55813886
#>  [62,]  0.50682601 -0.01371295
#>  [63,]  1.47970770 -0.47081586
#>  [64,] -1.45283859 -0.89249057
#>  [65,]  0.25273675 -0.01189409
#>  [66,]  1.62252696  0.82100355
#>  [67,] -0.33555943 -0.97862087
#>  [68,]  0.75599985 -1.24117527
#>  [69,] -1.51531108 -0.02748692
#>  [70,]  0.25043868 -0.26000958
#>  [71,] -0.91599443 -1.19892306
#>  [72,] -0.50609250  0.29325238
#>  [73,]  0.98596358 -1.64536814
#>  [74,] -1.74122769 -3.17086331
#>  [75,] -0.69179489 -2.03747540
#>  [76,]  1.10368119 -1.52772646
#>  [77,] -0.48761704 -0.82665250
#>  [78,] -0.18933512 -0.38903327
#>  [79,]  0.06786829 -0.16360421
#>  [80,] -0.93596828 -0.14060789
#>  [81,] -0.28199676 -0.81770904
#>  [82,] -0.10310374 -0.24114106
#>  [83,] -1.11398100 -1.95605237
#>  [84,]  1.58794327 -1.42848337
#>  [85,] -1.25368343 -1.92002197
#>  [86,]  1.84237364  2.56451926
#>  [87,]  1.02704030  0.06097657
#>  [88,]  0.91472883  0.23618593
#>  [89,]  0.83411789 -0.08784872
#>  [90,]  0.41631014 -1.61865022
#>  [91,]  0.14156347 -0.36766990
#>  [92,] -1.16814546  0.26880150
#>  [93,]  0.41692228  1.17324238
#>  [94,]  0.89965466 -0.96778162
#>  [95,] -0.23148008 -2.06181189
#>  [96,] -2.60895890  0.62655243
#>  [97,]  0.21118841 -1.42604391
#>  [98,] -1.00769805 -1.46323443
#>  [99,] -0.07931921  0.03213291
#> [100,] -0.60853559  0.15653988
#> 
#> $parameters
#>    alpha_1    alpha_2    beta_11    beta_21    beta_12    beta_22   gamma_11 
#>  0.9005992 -0.9498439  1.7562273  1.8362609  1.7978450  1.9471778  0.6002239 
#>   gamma_22   sigma_11   sigma_22 
#>  0.5047709  1.2892503  1.5614121 
#>