Quasi-hyperbolic Class

Class for the quasi-hyperbolic discounting model. Follows a similar structure to the model-class.

Slots

d

Integer denoting the number of dimensions of the model

k

Integer denoting the number of independent variables

n

Integer denoting the number of parameters

parameters

List containing the parameters relevant to a particular model instance

covariance

Numeric matrix denoting the residual covariance of the model

See also