Fit a model to data

This method enables the user to fit a model (defined in model) on a particular dataset (defined in data). The estimation procedure makes use of numerical optimization using either DEoptim or nloptr, as specified by the user. Estimation proceeds through an optimization according to the output of the objective function of the provided model, as defined through objective_function, thus using least-squares as optimization standard.

Usage

fit(model, data, ...)

# S4 method for class 'model,dataset'
fit(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "DEoptim",
  ...
)

Arguments

model

Instance of the model-class, defining the model to evaluate the objective function for.

data

Instance of the dataset-class containing the data to fit the model to.

...

Arguments passed on to the control parameters of the optimizer, either to DEoptim.control or the opts argument of nloptr.

dynamics

Character denoting the structure of the dynamical matrices. Can either be "anisotropic" (completely free), "symmetric" (symmetric around the diagonal), and "isotropic" (diagonal). Note that this influences different parameters for different models, namely \(\Gamma\) for the exponential discounting model, \(N\) and \(K\) for the quasi-hyperbolic discounting model, and \(\Gamma\) and \(N\) for the double-exponential discounting model. Defaults to "isotropic".

covariance

Character denoting the structure of covariance matrix. Can either by "symmetric" (symmetric around the diagonal) and "isotropic" (diagonal). Defaults to "symmetric".

optimizer

Character denoting the optimizer to use for the estimation. Can either be "DEoptim" for the differential evolution algorithm in DEoptim or "nloptr" for the library implemented in nloptr. Defaults to "DEoptim".

Value

An named list containing an instance of the model-class with the estimated parameters ("model"), the results of the optimization procedure ("fit"), the value of the objective after ending the optimization procedure ("objective"), the residuals of the model ("residuals"), and a named vector containing the values of the estimated parameters linked to a character vector explaining their content ("parameters").

Details

Note that currently, least-squares estimation is assumed, meaning that the covariance matrix is left out of the objective function. If maximum-likelihood were needed instead, then this function would need to change.

Examples

# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  itermax = 50,
  trace = FALSE
)
#> $model
#> Model of class "exponential":
#> 
#> Dimension: 2
#> Number of predictors: 2
#> Number of parameters: 10
#> 
#> Parameters:
#>   alpha: |  0.9668518  |
#>          |  -0.9117265  |
#> 
#>   beta: | 1.94535  1.950879 |
#>         | 1.810856  1.90557 |
#> 
#>   gamma: | 0.5645442  0.00 |
#>          | 0.00  0.5308806 |
#> 
#> 
#> Covariance: | 1.28104  0.00 |
#>             | 0.00  1.325322 |
#> 
#> $fit
#> $optim
#> $optim$bestmem
#>       par1       par2       par3       par4       par5       par6       par7 
#>  0.9668518 -0.9117265  1.9453497  1.8108556  1.9508795  1.9055696  0.5645442 
#>       par8 
#>  0.5308806 
#> 
#> $optim$bestval
#> [1] 260.7236
#> 
#> $optim$nfeval
#> [1] 4080
#> 
#> $optim$iter
#> [1] 50
#> 
#> 
#> $member
#> $member$lower
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>   -1   -1   -5   -5   -5   -5    0    0 
#> 
#> $member$upper
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>    1    1    5    5    5    5    1    1 
#> 
#> $member$bestmemit
#>           par1       par2     par3      par4     par5     par6      par7
#> 1  -0.27055693  0.4358981 2.170183 0.7127229 3.539568 0.352179 0.6440886
#> 2   0.41369192  0.1122859 2.020463 0.3258226 2.326234 3.274567 0.6691124
#> 3   0.41369192  0.1122859 2.020463 0.3258226 2.326234 3.274567 0.6691124
#> 4   0.08302356 -0.7083093 3.236261 1.5796185 1.995278 1.578788 0.1975439
#> 5   0.08302356 -0.7083093 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 6   0.08302356 -0.7083093 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 7   0.08302356 -0.6184576 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 8   0.08302356 -0.6184576 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 9   0.08302356 -0.6184576 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 10  0.08302356 -0.6184576 1.490071 1.5796185 1.995278 1.578788 0.1975439
#> 11  0.05191387 -0.2218408 1.770220 1.9602670 2.758788 1.876466 0.6630916
#> 12  0.05191387 -0.2218408 1.770220 1.9602670 2.758788 1.876466 0.6630916
#> 13  0.05191387 -0.2218408 1.770220 1.9602670 2.758788 1.876466 0.6630916
#> 14  0.81622461 -0.2218408 1.770220 1.9602670 2.758788 1.876466 0.6630916
#> 15  0.81622461 -0.2218408 1.770220 1.9602670 2.758788 1.876466 0.6630916
#> 16  0.50991755 -0.4607181 1.912682 2.1903499 2.577158 2.448932 0.4983552
#> 17  0.50991755 -0.4607181 1.912682 2.1903499 2.577158 2.448932 0.4983552
#> 18  0.50991755 -0.4607181 1.912682 2.1903499 2.577158 2.448932 0.4983552
#> 19  0.50991755 -0.4607181 1.912682 2.1903499 2.577158 2.448932 0.4983552
#> 20  0.77511628 -0.6227280 1.335867 1.4182950 1.636337 1.843788 0.6027381
#> 21  0.77511628 -0.6227280 1.335867 1.4182950 1.636337 1.881037 0.6353168
#> 22  0.77511628 -0.6227280 1.335867 1.4182950 1.636337 1.881037 0.6353168
#> 23  0.80250823 -0.7465223 1.849685 2.1818954 1.547027 1.735911 0.5496809
#> 24  0.80250823 -0.7465223 1.849685 2.1818954 1.547027 1.735911 0.5496809
#> 25  0.80250823 -0.7465223 1.849685 2.1818954 1.547027 1.735911 0.5496809
#> 26  0.80250823 -0.7465223 1.849685 2.1818954 1.547027 1.735911 0.5496809
#> 27  0.96858981 -0.7212151 1.829344 2.0064338 2.354785 1.721079 0.5115945
#> 28  0.96858981 -0.7212151 1.829344 2.0064338 2.354785 1.721079 0.5115945
#> 29  0.96858981 -0.7212151 1.829344 2.0064338 2.354785 1.721079 0.5115945
#> 30  0.80250823 -0.7465223 1.849685 1.9667419 1.860830 1.735911 0.5496809
#> 31  0.95475870 -0.8927698 2.170183 1.4629091 2.119593 1.892341 0.5466044
#> 32  0.95475870 -0.8927698 2.170183 1.4629091 2.119593 1.892341 0.5466044
#> 33  0.95475870 -0.8927698 1.623570 1.5184599 2.024742 1.892341 0.5466044
#> 34  0.95475870 -0.8927698 1.623570 1.5184599 2.024742 1.892341 0.5466044
#> 35  0.95475870 -0.8927698 1.623570 1.5184599 2.024742 1.892341 0.5466044
#> 36  0.95475870 -0.8927698 1.623570 1.5184599 2.024742 1.892341 0.5466044
#> 37  0.95475870 -0.8927698 1.623570 1.5184599 2.024742 1.892341 0.5466044
#> 38  0.95475870 -0.8927698 2.086303 1.5184599 2.024742 1.892341 0.5466044
#> 39  0.95475870 -0.8927698 2.086303 1.5184599 2.024742 1.892341 0.5466044
#> 40  0.95475870 -0.8160601 2.086303 1.5184599 2.024742 1.892341 0.5466044
#> 41  0.89254196 -0.6851346 1.812970 1.6941536 1.779537 1.966440 0.5603323
#> 42  0.89254196 -0.6851346 1.812970 1.6941536 1.779537 1.966440 0.5603323
#> 43  0.89254196 -0.6851346 1.812970 1.6941536 1.779537 1.966440 0.5603323
#> 44  0.89254196 -0.6851346 1.812970 1.6941536 1.779537 1.966440 0.5603323
#> 45  0.89254196 -0.6851346 1.812970 1.6941536 1.779537 1.966440 0.5603323
#> 46  0.96685180 -0.7221627 1.770220 1.7969063 2.152306 1.950705 0.5645442
#> 47  0.96685180 -0.9117265 1.945350 1.7969063 2.152306 1.950705 0.5645442
#> 48  0.96685180 -0.9117265 1.945350 1.8108556 1.950879 2.058618 0.5645442
#> 49  0.96685180 -0.9117265 1.945350 1.8108556 1.950879 2.058618 0.5645442
#> 50  0.96685180 -0.9117265 1.945350 1.8108556 1.950879 2.058618 0.5645442
#>          par8
#> 1  0.04417196
#> 2  0.18050158
#> 3  0.18050158
#> 4  0.45964743
#> 5  0.45964743
#> 6  0.45964743
#> 7  0.45964743
#> 8  0.45964743
#> 9  0.45964743
#> 10 0.51287416
#> 11 0.30158835
#> 12 0.30158835
#> 13 0.30158835
#> 14 0.30158835
#> 15 0.30158835
#> 16 0.34222021
#> 17 0.34222021
#> 18 0.34222021
#> 19 0.34222021
#> 20 0.39245746
#> 21 0.39245746
#> 22 0.39245746
#> 23 0.48386858
#> 24 0.48386858
#> 25 0.48386858
#> 26 0.48386858
#> 27 0.40618117
#> 28 0.40618117
#> 29 0.40618117
#> 30 0.48386858
#> 31 0.52909369
#> 32 0.52909369
#> 33 0.52909369
#> 34 0.52909369
#> 35 0.52909369
#> 36 0.52909369
#> 37 0.52909369
#> 38 0.52909369
#> 39 0.52909369
#> 40 0.52909369
#> 41 0.53020075
#> 42 0.53020075
#> 43 0.53020075
#> 44 0.53020075
#> 45 0.53020075
#> 46 0.53088060
#> 47 0.53088060
#> 48 0.53088060
#> 49 0.53088060
#> 50 0.53088060
#> 
#> $member$bestvalit
#>  [1] 1116.1882  750.3105  750.3105  696.3594  566.5022  566.5022  565.6210
#>  [8]  565.6210  565.6210  560.9113  472.4826  472.4826  472.4826  444.7874
#> [15]  444.7874  386.9721  386.9721  386.9721  386.9721  332.3286  327.5785
#> [22]  327.5785  318.7635  318.7635  318.7635  318.7635  298.9751  298.9751
#> [29]  298.9751  280.9091  280.7498  280.7498  273.1246  273.1246  273.1246
#> [36]  273.1246  273.1246  270.6984  270.6984  269.3405  267.1382  267.1382
#> [43]  267.1382  267.1382  267.1382  266.5651  265.5811  261.9930  261.9930
#> [50]  261.9930
#> 
#> $member$pop
#>            [,1]       [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.9668518 -0.9117265 1.945350 1.810856 1.950879 1.905570 0.5645442
#>  [2,] 0.8767539 -0.8421020 1.763014 1.804291 1.971733 1.799176 0.5533891
#>  [3,] 0.8956744 -0.6617906 1.914387 1.907702 2.068635 2.076991 0.5675571
#>  [4,] 0.9947843 -0.5946549 1.872622 1.601936 2.370509 2.003687 0.3404521
#>  [5,] 0.9453532 -0.9607759 1.769766 1.754573 2.058123 1.923903 0.5769551
#>  [6,] 0.9704707 -0.7781492 2.141981 2.112956 2.160607 1.692052 0.5210682
#>  [7,] 0.9015522 -0.8586393 2.041141 1.441851 1.958003 2.064804 0.5076728
#>  [8,] 0.8036044 -0.8946112 1.949256 1.852921 1.814518 2.103906 0.5947195
#>  [9,] 0.9694349 -0.5152271 1.905847 1.761834 2.064225 1.475000 0.5038442
#> [10,] 0.9185492 -0.8462973 1.136172 2.294118 1.678294 1.977838 0.6453123
#> [11,] 0.9070693 -0.6477607 1.635572 1.923144 1.998202 2.040151 0.5694761
#> [12,] 0.7962613 -0.4440404 2.085105 1.782692 2.142241 2.110303 0.5565777
#> [13,] 0.8771666 -0.4751716 1.940134 1.686533 1.916477 1.841692 0.5032679
#> [14,] 0.8988371 -0.5532733 1.749607 1.732053 1.892197 2.018982 0.5565515
#> [15,] 0.9746192 -0.6945040 2.129969 1.401360 2.195438 1.915513 0.5563109
#> [16,] 0.9144426 -0.7895882 2.033314 1.709611 2.083545 2.066890 0.5060299
#> [17,] 0.9946231 -0.9972775 1.917725 1.796715 1.854290 2.096710 0.5888041
#> [18,] 0.8129318 -0.9051264 1.840500 1.597076 2.034372 1.853406 0.5697725
#> [19,] 0.9505299 -0.9277514 2.025848 1.781403 1.622169 1.888575 0.5584495
#> [20,] 0.8034386 -0.8125791 1.768484 1.958469 1.748682 2.553494 0.6250184
#> [21,] 0.7469928 -0.8383479 2.054312 1.535566 1.878318 1.918280 0.5861036
#> [22,] 0.9018358 -0.4236559 2.129635 1.642308 2.090989 1.959691 0.5942248
#> [23,] 0.8906137 -0.5790834 1.482227 2.125316 2.191380 2.312661 0.5912329
#> [24,] 0.9685898 -0.6182356 1.829344 1.574229 1.829493 1.890350 0.5115945
#> [25,] 0.9581242 -0.7676919 1.854645 1.401974 1.848590 1.796461 0.5326658
#> [26,] 0.9914389 -0.8922816 2.243688 1.751659 2.012257 1.929934 0.5502302
#> [27,] 0.9070149 -0.6923501 1.658795 1.775706 1.584509 1.712330 0.6859817
#> [28,] 0.8963029 -0.7782758 2.047990 1.233409 2.043061 2.157861 0.3694069
#> [29,] 0.9631851 -0.2632085 1.813404 1.626849 1.764919 1.893564 0.6123150
#> [30,] 0.9396995 -0.8798810 1.566048 1.640175 1.986757 1.881037 0.5928131
#> [31,] 0.7997087 -0.6081068 1.935655 1.985686 2.083265 1.729355 0.5655603
#> [32,] 0.4554769 -0.6389145 1.999064 1.712616 1.901372 2.008187 0.6155359
#> [33,] 0.7159728 -0.7173732 1.909906 1.290955 2.143218 1.999382 0.5859727
#> [34,] 0.9460437 -0.8851940 1.983766 1.890182 2.467763 1.871649 0.3826510
#> [35,] 0.9011932 -0.9696315 2.016058 1.666986 1.931349 2.113328 0.5488454
#> [36,] 0.9875810 -0.9630074 2.008731 1.780721 1.935619 1.878219 0.5499216
#> [37,] 0.9872137 -0.8248082 1.708381 1.742113 1.995278 2.031885 0.6067632
#> [38,] 0.8925420 -0.6860317 1.812970 1.694154 1.779537 1.958724 0.5603323
#> [39,] 0.9413862 -0.8569571 1.892724 1.726767 1.784273 1.729647 0.5368599
#> [40,] 0.8334055 -0.9192064 1.759674 1.965547 1.560877 2.209986 0.6513563
#> [41,] 0.9656616 -0.6220046 2.159583 1.727664 2.119355 1.981579 0.5477893
#> [42,] 0.8815291 -0.6896114 1.750254 1.579268 1.509230 2.203447 0.6394499
#> [43,] 0.8845322 -0.9511393 1.826377 2.010482 1.766920 1.896440 0.5728284
#> [44,] 0.7625281 -0.5367170 1.940022 1.552572 1.974292 1.919853 0.5791534
#> [45,] 0.8952083 -0.8540044 2.183954 1.343834 1.999371 1.817680 0.6025485
#> [46,] 0.7958165 -0.8268501 1.897867 1.948966 1.910510 1.896571 0.5696702
#> [47,] 0.7953941 -0.8192972 1.757631 1.823622 2.038663 1.870782 0.5973348
#> [48,] 0.9802280 -0.9187656 1.892469 1.679780 2.060873 2.310803 0.4853270
#> [49,] 0.8307751 -0.6471701 2.086042 1.744511 1.800616 2.075536 0.6098066
#> [50,] 0.9529520 -0.7118278 1.703491 1.885028 1.708210 1.957302 0.6406334
#> [51,] 0.9187281 -0.9959142 1.626877 1.611362 1.726033 1.721390 0.5959902
#> [52,] 0.9342234 -0.6142286 2.190261 1.289779 1.891432 1.882600 0.5551676
#> [53,] 0.9478989 -0.6804105 1.833350 1.682463 2.033924 1.961332 0.4949987
#> [54,] 0.9244115 -0.7804927 1.771319 1.702528 1.888660 2.065557 0.5189160
#> [55,] 0.9695260 -0.8179443 1.638109 1.636852 1.533898 2.401323 0.6568719
#> [56,] 0.7001495 -0.8356858 1.697552 1.732791 2.190294 1.958837 0.5764374
#> [57,] 0.9036281 -0.3588762 1.824500 1.826288 2.130408 2.002620 0.2928600
#> [58,] 0.7154378 -0.7448884 1.854974 1.615716 1.911679 1.958941 0.5033426
#> [59,] 0.9766396 -0.8070075 1.933594 1.503689 1.752272 1.763620 0.5750412
#> [60,] 0.9651172 -0.4886790 2.210398 1.741417 1.834598 1.876645 0.4815207
#> [61,] 0.8237103 -0.6270886 2.072862 1.689269 1.965446 2.048938 0.6404013
#> [62,] 0.9523305 -0.9608576 1.537527 1.954494 2.308282 2.167345 0.4621655
#> [63,] 0.9699353 -0.5819914 1.905749 1.706566 1.802952 1.909838 0.4810834
#> [64,] 0.9694790 -0.6614199 2.020880 1.794208 2.011803 2.120376 0.5513922
#> [65,] 0.6477022 -0.8124105 1.781003 1.465916 1.962816 2.091695 0.6301430
#> [66,] 0.9465776 -0.6903033 2.255394 1.672471 1.991213 2.019697 0.5405600
#> [67,] 0.9317069 -0.8964463 1.930791 1.865399 2.015239 2.099858 0.6704584
#> [68,] 0.9947662 -0.6952627 1.663695 1.897068 2.101584 1.957820 0.5750215
#> [69,] 0.9606162 -0.6624514 1.943741 1.728006 1.933557 2.052369 0.5948778
#> [70,] 0.9475704 -0.7435930 1.542139 1.874161 2.087908 1.924971 0.5699216
#> [71,] 0.8899472 -0.6957881 1.814233 1.633474 2.079133 1.837010 0.5665455
#> [72,] 0.9817609 -0.5775624 1.892086 1.821253 1.964847 2.117509 0.5152041
#> [73,] 0.9143561 -0.9959636 1.877044 1.916789 2.346962 2.010049 0.5848684
#> [74,] 0.9636518 -0.6947644 1.782969 1.473827 1.907439 2.197640 0.6343428
#> [75,] 0.8650863 -0.8459725 1.675447 1.513055 1.855820 1.797755 0.5763194
#> [76,] 0.9577123 -0.9180192 1.510336 1.690124 2.135248 1.999414 0.6211532
#> [77,] 0.9622087 -0.6258094 2.229195 1.684643 2.005941 2.252067 0.5040070
#> [78,] 0.9234858 -0.7586549 1.824088 1.470955 2.031039 1.983100 0.5028852
#> [79,] 0.9914612 -0.8967524 2.123236 1.923101 1.985104 1.992565 0.5921444
#> [80,] 0.9547587 -0.8160601 2.086303 1.518460 2.010552 1.892341 0.5466044
#>            [,8]
#>  [1,] 0.5308806
#>  [2,] 0.5819674
#>  [3,] 0.4181720
#>  [4,] 0.4302508
#>  [5,] 0.5389521
#>  [6,] 0.4719853
#>  [7,] 0.4924140
#>  [8,] 0.4789325
#>  [9,] 0.5276087
#> [10,] 0.4340532
#> [11,] 0.4995407
#> [12,] 0.4918747
#> [13,] 0.4800066
#> [14,] 0.4169329
#> [15,] 0.5252301
#> [16,] 0.5741680
#> [17,] 0.5665499
#> [18,] 0.5519980
#> [19,] 0.5710449
#> [20,] 0.4096439
#> [21,] 0.5783770
#> [22,] 0.2846682
#> [23,] 0.4676040
#> [24,] 0.4725560
#> [25,] 0.5631899
#> [26,] 0.5807973
#> [27,] 0.5303203
#> [28,] 0.4735243
#> [29,] 0.4246094
#> [30,] 0.6047602
#> [31,] 0.5465522
#> [32,] 0.4651047
#> [33,] 0.5524774
#> [34,] 0.6262279
#> [35,] 0.5265657
#> [36,] 0.5194668
#> [37,] 0.4582457
#> [38,] 0.5302008
#> [39,] 0.5498739
#> [40,] 0.4921236
#> [41,] 0.4606899
#> [42,] 0.2835335
#> [43,] 0.5209707
#> [44,] 0.4764441
#> [45,] 0.5475946
#> [46,] 0.5175550
#> [47,] 0.5136047
#> [48,] 0.4902188
#> [49,] 0.5090501
#> [50,] 0.5196851
#> [51,] 0.5103129
#> [52,] 0.4767197
#> [53,] 0.4185717
#> [54,] 0.5033914
#> [55,] 0.3885523
#> [56,] 0.4805696
#> [57,] 0.5457118
#> [58,] 0.4836653
#> [59,] 0.5950536
#> [60,] 0.4771158
#> [61,] 0.5081008
#> [62,] 0.5384634
#> [63,] 0.4123033
#> [64,] 0.5039895
#> [65,] 0.5162911
#> [66,] 0.4870269
#> [67,] 0.5273344
#> [68,] 0.4821480
#> [69,] 0.4527815
#> [70,] 0.5042060
#> [71,] 0.4629968
#> [72,] 0.2869068
#> [73,] 0.4621923
#> [74,] 0.5049150
#> [75,] 0.5884321
#> [76,] 0.5383735
#> [77,] 0.4700242
#> [78,] 0.4449498
#> [79,] 0.5371482
#> [80,] 0.5290937
#> 
#> $member$storepop
#> list()
#> 
#> 
#> attr(,"class")
#> [1] "DEoptim"
#> 
#> $objective
#> [1] 260.7236
#> 
#> $residuals
#>               [,1]         [,2]
#>   [1,] -0.41800950 -0.362917193
#>   [2,] -0.09091349 -0.030933639
#>   [3,] -0.05844108  0.300852212
#>   [4,]  1.34614616 -0.560873011
#>   [5,] -1.88409298 -2.187926581
#>   [6,] -0.28918816  0.517091543
#>   [7,]  0.03650522 -0.017128629
#>   [8,] -1.25991580 -1.694483991
#>   [9,]  0.25136960  0.595349590
#>  [10,] -0.80512734  0.798774314
#>  [11,] -1.32670009 -3.601380100
#>  [12,] -0.38444622 -0.288497212
#>  [13,]  0.92103430  0.927814462
#>  [14,]  2.05230812  1.872574476
#>  [15,]  1.65413187 -0.079902449
#>  [16,]  2.01698203  0.388360699
#>  [17,]  0.26380115 -0.624270694
#>  [18,]  0.55767839 -0.822715775
#>  [19,] -0.01032297 -0.346900323
#>  [20,] -0.85344780  0.237099742
#>  [21,]  0.27515060 -0.930177505
#>  [22,] -0.83972580  0.497424686
#>  [23,] -1.08342814 -0.801097105
#>  [24,]  0.16271663  2.063310866
#>  [25,]  1.26156645  1.684765527
#>  [26,]  1.32311684 -0.799114827
#>  [27,] -0.76633223 -1.328597991
#>  [28,] -1.55954706  0.144107872
#>  [29,] -1.51216764 -1.092564868
#>  [30,]  0.68228366 -0.869579653
#>  [31,]  0.02348202 -0.096633198
#>  [32,]  0.31960060  1.888741182
#>  [33,]  2.53474898  2.168158910
#>  [34,] -0.20286920 -0.885064483
#>  [35,]  1.90260657  0.663786270
#>  [36,]  0.24426558  1.123421185
#>  [37,] -0.90962008  2.349064118
#>  [38,]  1.88921445  0.685980209
#>  [39,]  0.11422377  0.678097897
#>  [40,]  0.99937855 -0.862078723
#>  [41,]  0.55215629  1.180475277
#>  [42,] -0.74168028 -1.238400430
#>  [43,] -1.63404458 -1.178593005
#>  [44,]  0.91577110 -0.969843990
#>  [45,]  0.77143681  0.341828457
#>  [46,] -0.07836318 -0.441917127
#>  [47,]  0.34947598  0.617339451
#>  [48,] -0.81231060 -0.334368517
#>  [49,] -0.17957386  1.487517697
#>  [50,]  3.58296354 -0.398613218
#>  [51,] -0.40353032  2.842060042
#>  [52,]  1.23991944  0.195223921
#>  [53,]  0.34777020  0.659107371
#>  [54,] -1.61024182  1.071833402
#>  [55,]  0.91341434  0.976911290
#>  [56,]  0.76823104 -0.523849005
#>  [57,]  0.54750896 -0.930724498
#>  [58,]  1.46294618  1.421353708
#>  [59,] -1.68451895  0.160672804
#>  [60,] -0.95707253 -1.299582229
#>  [61,] -1.28168936 -1.833698215
#>  [62,]  1.91692874  2.612240201
#>  [63,]  0.17786493 -1.679159130
#>  [64,] -0.70122526  0.072001940
#>  [65,] -0.28516741  0.363649821
#>  [66,] -1.23896391 -1.071371312
#>  [67,] -0.26246293  0.311806766
#>  [68,] -1.30609483 -0.776541418
#>  [69,]  0.71517481  1.324236086
#>  [70,]  0.23140925  0.870567021
#>  [71,]  1.23889211  1.687394016
#>  [72,]  0.17199138 -0.656998772
#>  [73,] -0.23552333  0.384240500
#>  [74,]  0.26487949 -0.326911065
#>  [75,] -2.01237906 -1.538413314
#>  [76,] -1.18372875  0.284096160
#>  [77,] -0.04735258  0.800795893
#>  [78,]  1.51736528  0.337716688
#>  [79,]  0.12047455 -0.401144773
#>  [80,] -0.55399985 -0.959167397
#>  [81,] -0.28900320 -0.116963934
#>  [82,]  3.60081118 -0.493327585
#>  [83,] -0.36077191  0.564068509
#>  [84,]  0.80266778 -0.365425949
#>  [85,]  0.38690483  2.025906783
#>  [86,] -1.02052852 -1.451360631
#>  [87,]  0.58982339 -1.258223585
#>  [88,] -0.37140245  0.423939681
#>  [89,]  0.93009879  0.007116071
#>  [90,]  1.11185490 -0.046469165
#>  [91,]  1.04894818  0.891388604
#>  [92,]  1.29137895  3.163877249
#>  [93,] -0.13089875 -0.170623474
#>  [94,]  0.32277245  0.968912593
#>  [95,] -1.90525141 -0.688216517
#>  [96,] -0.22142888  0.277044628
#>  [97,]  0.21667310  0.629115605
#>  [98,] -1.43023151  1.080233250
#>  [99,]  0.09806116  0.095695229
#> [100,]  1.60677774 -0.879058215
#> 
#> $parameters
#>    alpha_1    alpha_2    beta_11    beta_21    beta_12    beta_22   gamma_11 
#>  0.9668518 -0.9117265  1.9453497  1.8108556  1.9508795  1.9055696  0.5645442 
#>   gamma_22   sigma_11   sigma_22 
#>  0.5308806  1.2810397  1.3253222 
#>