Fit a model to data

This method enables the user to fit a model (defined in model) on a particular dataset (defined in data). The estimation procedure makes use of numerical optimization using either DEoptim or nloptr, as specified by the user. Estimation proceeds through an optimization according to the output of the objective function of the provided model, as defined through objective_function, thus using least-squares as optimization standard.

Usage

fit(model, data, ...)

# S4 method for class 'model,dataset'
fit(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "DEoptim",
  ...
)

Arguments

model

Instance of the model-class, defining the model to evaluate the objective function for.

data

Instance of the dataset-class containing the data to fit the model to.

...

Arguments passed on to the control parameters of the optimizer, either to DEoptim.control or the opts argument of nloptr.

dynamics

Character denoting the structure of the dynamical matrices. Can either be "anisotropic" (completely free), "symmetric" (symmetric around the diagonal), and "isotropic" (diagonal). Note that this influences different parameters for different models, namely \(\Gamma\) for the exponential discounting model, \(N\) and \(K\) for the quasi-hyperbolic discounting model, and \(\Gamma\) and \(N\) for the double-exponential discounting model. Defaults to "isotropic".

covariance

Character denoting the structure of covariance matrix. Can either by "symmetric" (symmetric around the diagonal) and "isotropic" (diagonal). Defaults to "symmetric".

optimizer

Character denoting the optimizer to use for the estimation. Can either be "DEoptim" for the differential evolution algorithm in DEoptim or "nloptr" for the library implemented in nloptr. Defaults to "DEoptim".

Value

An named list containing an instance of the model-class with the estimated parameters ("model"), the results of the optimization procedure ("fit"), the value of the objective after ending the optimization procedure ("objective"), the residuals of the model ("residuals"), and a named vector containing the values of the estimated parameters linked to a character vector explaining their content ("parameters").

Details

Note that currently, least-squares estimation is assumed, meaning that the covariance matrix is left out of the objective function. If maximum-likelihood were needed instead, then this function would need to change.

Examples

# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  itermax = 50,
  trace = FALSE
)
#> $model
#> Model of class "exponential":
#> 
#> Dimension: 2
#> Number of predictors: 2
#> Number of parameters: 10
#> 
#> Parameters:
#>   alpha: |  0.977926  |
#>          |  -0.7421743  |
#> 
#>   beta: | 1.894378  1.942535 |
#>         | 1.805858  1.930499 |
#> 
#>   gamma: | 0.5034501  0.00 |
#>          | 0.00  0.5053175 |
#> 
#> 
#> Covariance: | 1.215134  0.00 |
#>             | 0.00  1.265273 |
#> 
#> $fit
#> $optim
#> $optim$bestmem
#>       par1       par2       par3       par4       par5       par6       par7 
#>  0.9779260 -0.7421743  1.8943784  1.8058577  1.9425347  1.9304992  0.5034501 
#>       par8 
#>  0.5053175 
#> 
#> $optim$bestval
#> [1] 251.8245
#> 
#> $optim$nfeval
#> [1] 4080
#> 
#> $optim$iter
#> [1] 50
#> 
#> 
#> $member
#> $member$lower
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>   -1   -1   -5   -5   -5   -5    0    0 
#> 
#> $member$upper
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>    1    1    5    5    5    5    1    1 
#> 
#> $member$bestmemit
#>           par1       par2     par3      par4     par5     par6      par7
#> 1  -0.27055693  0.4358981 2.170183 0.7127229 3.539568 0.352179 0.6440886
#> 2   0.48317140  0.2951388 1.925554 2.6202437 1.758339 3.285307 0.1470606
#> 3   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 4   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 5   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 6   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 7   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 8   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 9   0.94787645 -0.5094947 1.794533 3.1533020 2.583137 1.751094 0.1201802
#> 10  0.02520247 -0.9497438 1.697248 1.0700275 1.968165 2.211439 0.7959482
#> 11  0.62888428 -0.5862240 2.611697 1.4550789 2.120595 2.135551 0.1488422
#> 12  0.62888428 -0.5862240 2.611697 1.4550789 2.120595 2.135551 0.1488422
#> 13  0.70212881 -0.9497438 1.697248 1.0700275 1.968165 2.211439 0.7959482
#> 14  0.70212881 -0.9497438 1.697248 1.0700275 1.968165 2.211439 0.7959482
#> 15  0.97997134 -0.7880692 2.335892 1.7118055 2.387062 1.934041 0.1414340
#> 16  0.97997134 -0.7880692 2.335892 1.7118055 2.387062 1.934041 0.1414340
#> 17  0.73259123 -0.4345720 1.390356 1.5387918 1.700782 1.956733 0.5106950
#> 18  0.73259123 -0.4345720 1.587848 1.5387918 1.700782 1.956733 0.5106950
#> 19  0.73259123 -0.4345720 1.587848 1.5387918 1.700782 1.956733 0.5696562
#> 20  0.97997134 -0.7880692 2.335892 1.7118055 2.387062 1.934041 0.4539951
#> 21  0.97997134 -0.7880692 2.335892 1.7118055 2.387062 1.934041 0.4539951
#> 22  0.97997134 -0.7880692 2.335892 1.7118055 2.387062 1.934041 0.4539951
#> 23  0.85178417 -0.8730142 2.053227 1.6881465 1.605573 2.192321 0.6189004
#> 24  0.85178417 -0.8730142 2.053227 1.6881465 1.605573 2.192321 0.6189004
#> 25  0.85178417 -0.8730142 2.053227 1.6881465 1.605573 2.192321 0.6189004
#> 26  0.85178417 -0.8730142 2.053227 1.6881465 1.605573 2.192321 0.6189004
#> 27  0.85178417 -0.8730142 2.053227 1.6881465 1.605573 2.192321 0.6189004
#> 28  0.92496495 -0.9268970 1.525856 1.7602763 1.781401 2.162652 0.6289422
#> 29  0.92496495 -0.9268970 1.525856 1.7602763 1.781401 2.162652 0.6289422
#> 30  0.92496495 -0.9563477 1.525856 1.7602763 1.781401 2.162652 0.6289422
#> 31  0.73259123 -0.7121059 1.957234 1.5387918 2.024738 1.956733 0.5696562
#> 32  0.91098748 -0.5968072 1.855472 1.8630619 1.770662 2.012634 0.5829996
#> 33  0.91098748 -0.5968072 1.855472 1.8630619 1.770662 2.012634 0.5829996
#> 34  0.91098748 -0.5968072 1.855472 1.8630619 1.770662 2.012634 0.5829996
#> 35  0.85178417 -0.8730142 2.053227 1.6881465 1.965431 2.192321 0.6189004
#> 36  0.85178417 -0.8730142 2.053227 1.6881465 1.965431 2.192321 0.6189004
#> 37  0.96725749 -0.8730142 2.053227 1.6881465 1.965431 2.192321 0.6189004
#> 38  0.86662288 -0.7999049 1.995047 1.7867028 2.205390 2.034880 0.4721870
#> 39  0.86662288 -0.7999049 1.995047 1.7867028 2.205390 2.034880 0.4721870
#> 40  0.96725749 -0.8730142 2.053227 1.6881465 1.965431 2.140071 0.5669146
#> 41  0.96725749 -0.8730142 2.053227 1.6881465 1.965431 2.140071 0.5669146
#> 42  0.96725749 -0.8730142 2.053227 1.6881465 1.965431 2.140071 0.5669146
#> 43  0.96725749 -0.8730142 2.053227 1.6881465 1.965431 2.140071 0.5669146
#> 44  0.96725749 -0.8730142 1.781102 1.6881465 1.965431 2.140071 0.5669146
#> 45  0.96725749 -0.8730142 1.781102 1.6881465 1.965431 2.140071 0.5669146
#> 46  0.96725749 -0.8730142 1.781102 1.6881465 1.965431 2.140071 0.5669146
#> 47  0.96725749 -0.8730142 1.781102 1.8107339 1.965431 2.140071 0.5669146
#> 48  0.96725749 -0.8730142 1.781102 1.8107339 1.965431 2.140071 0.5669146
#> 49  0.91098748 -0.8944437 1.977593 1.7310177 1.989219 1.921683 0.5090669
#> 50  0.99586938 -0.8302088 2.114864 1.6542730 1.937890 1.950068 0.5425766
#>          par8
#> 1  0.04417196
#> 2  0.45499081
#> 3  0.20680428
#> 4  0.20680428
#> 5  0.20680428
#> 6  0.20680428
#> 7  0.20680428
#> 8  0.20680428
#> 9  0.20680428
#> 10 0.46198100
#> 11 0.40119611
#> 12 0.40119611
#> 13 0.46198100
#> 14 0.46198100
#> 15 0.42118930
#> 16 0.42118930
#> 17 0.23773067
#> 18 0.23773067
#> 19 0.23773067
#> 20 0.42118930
#> 21 0.42118930
#> 22 0.62978645
#> 23 0.52700351
#> 24 0.52700351
#> 25 0.52700351
#> 26 0.52700351
#> 27 0.52700351
#> 28 0.59411454
#> 29 0.59411454
#> 30 0.59411454
#> 31 0.57275943
#> 32 0.58729695
#> 33 0.58729695
#> 34 0.58729695
#> 35 0.52700351
#> 36 0.52700351
#> 37 0.56578616
#> 38 0.56244900
#> 39 0.56244900
#> 40 0.56578616
#> 41 0.56578616
#> 42 0.56578616
#> 43 0.56578616
#> 44 0.56578616
#> 45 0.56578616
#> 46 0.56578616
#> 47 0.56578616
#> 48 0.56578616
#> 49 0.59368033
#> 50 0.54528206
#> 
#> $member$bestvalit
#>  [1] 1186.5569  804.0266  594.3962  594.3962  594.3962  594.3962  594.3962
#>  [8]  594.3962  594.3962  450.5874  447.9401  447.9401  431.0912  431.0912
#> [15]  372.5902  372.5902  354.5748  339.7669  330.1174  291.0099  291.0099
#> [22]  290.9169  279.2086  279.2086  279.2086  279.2086  279.2086  273.8262
#> [29]  273.8262  273.6458  269.2872  268.2574  268.2574  268.2574  267.1350
#> [36]  267.1350  267.0848  265.8281  265.8281  257.8993  257.8993  257.8993
#> [43]  257.8993  255.8297  255.8297  255.8297  254.7931  254.7931  254.7606
#> [50]  252.9453
#> 
#> $member$pop
#>            [,1]       [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.9912047 -0.9755344 1.888304 1.722731 2.010500 1.885955 0.6525328
#>  [2,] 0.9687716 -0.9826544 1.922658 1.635721 1.992902 2.085174 0.5728964
#>  [3,] 0.8532261 -0.9155695 1.792139 1.814288 1.811521 2.193712 0.6708937
#>  [4,] 0.9323505 -0.8139167 1.945052 1.835437 2.126669 2.398569 0.5235937
#>  [5,] 0.9245820 -0.9695588 1.703658 1.500449 1.996791 1.778112 0.4556411
#>  [6,] 0.8576104 -0.8374740 1.915361 1.621556 1.951101 1.669247 0.5671757
#>  [7,] 0.8307663 -0.7153313 1.768863 1.791228 1.847566 1.884576 0.5765481
#>  [8,] 0.9992070 -0.7997400 1.963268 1.789907 2.022704 1.900312 0.5562508
#>  [9,] 0.7999461 -0.8394613 1.845856 1.760837 1.532059 2.009502 0.6460020
#> [10,] 0.8027656 -0.5149492 2.000753 1.670780 1.920862 1.792869 0.6487125
#> [11,] 0.9041730 -0.6568904 1.991516 1.790956 1.919310 2.160460 0.6006198
#> [12,] 0.9136987 -0.8216208 1.811883 1.993464 1.842944 2.090958 0.6314676
#> [13,] 0.8340853 -0.5471419 1.946619 2.029343 2.076996 1.914138 0.5387627
#> [14,] 0.9672575 -0.8730142 1.781102 1.810734 1.965431 2.140071 0.5669146
#> [15,] 0.9532979 -0.8289328 1.633279 2.132526 2.208287 2.125464 0.5290678
#> [16,] 0.9875810 -0.8482551 1.911820 1.522625 1.795783 1.357604 0.4066352
#> [17,] 0.8831393 -0.8398447 1.411807 1.718045 1.687190 1.685901 0.6189036
#> [18,] 0.9995757 -0.8988859 1.809238 1.356387 2.233926 1.519688 0.4918087
#> [19,] 0.8887246 -0.7011365 1.984975 1.927085 2.140795 2.097889 0.6085236
#> [20,] 0.8777766 -0.7884022 2.023967 1.714212 1.868161 2.178952 0.6903981
#> [21,] 0.8517160 -0.8586670 1.711408 1.897554 2.235158 2.026764 0.4329332
#> [22,] 0.9608413 -0.9362977 1.925554 2.421115 1.936603 1.714170 0.6444182
#> [23,] 0.9779260 -0.7421743 1.894378 1.805858 1.942535 1.930499 0.5034501
#> [24,] 0.9950260 -0.6456355 2.158046 2.047817 2.270890 2.001324 0.4865606
#> [25,] 0.9847123 -0.7506390 1.233283 1.855511 2.130558 1.899208 0.5944342
#> [26,] 0.7992692 -0.8758193 1.754453 1.814636 1.844861 1.995906 0.5004608
#> [27,] 0.9110678 -0.7046818 1.839662 2.060011 2.329342 1.904744 0.4325787
#> [28,] 0.9426396 -0.9740953 2.054410 1.844033 1.830800 1.850033 0.5348976
#> [29,] 0.9691629 -0.7731843 1.810500 1.316206 1.925372 1.747038 0.5037429
#> [30,] 0.8951535 -0.8389608 1.888015 1.730279 2.341898 1.904436 0.3281093
#> [31,] 0.9604928 -0.5963814 1.811668 1.481027 1.723935 1.881524 0.5465862
#> [32,] 0.9378038 -0.8831853 1.722396 1.549350 1.793843 2.031389 0.6350491
#> [33,] 0.9891817 -0.8275087 1.538390 1.645941 1.882411 1.833037 0.5755465
#> [34,] 0.9748447 -0.9206251 1.803700 1.483414 2.194578 1.809709 0.5725883
#> [35,] 0.9259750 -0.9961524 1.832012 2.173839 1.877197 1.884351 0.5612544
#> [36,] 0.9061074 -0.7069112 1.760784 1.790278 1.835747 1.781779 0.5571617
#> [37,] 0.9804120 -0.7730786 1.867682 1.758124 1.790061 1.687407 0.5277792
#> [38,] 0.9518957 -0.7212854 1.791541 1.944198 1.702160 1.904806 0.5310321
#> [39,] 0.9313948 -0.8799170 1.995509 1.639275 2.181261 1.898432 0.5188515
#> [40,] 0.9904926 -0.8935543 1.912925 1.444870 1.867440 1.954289 0.5937131
#> [41,] 0.9253255 -0.4313436 1.780273 2.071971 2.018445 1.944440 0.5479303
#> [42,] 0.8666229 -0.7999049 1.873295 1.854846 2.021543 1.860075 0.5697317
#> [43,] 0.9707776 -0.6676247 1.820743 1.485338 2.039212 1.966550 0.5617911
#> [44,] 0.9249649 -0.9563477 1.525856 1.760276 1.781401 2.162652 0.5840423
#> [45,] 0.8766413 -0.8721938 1.744053 1.493057 2.192336 2.035689 0.5610746
#> [46,] 0.9412706 -0.9189651 1.574108 1.829372 2.289439 1.930259 0.5809300
#> [47,] 0.9438375 -0.9110711 2.020581 1.653064 2.345964 1.933997 0.5328344
#> [48,] 0.9958694 -0.8302088 2.114864 1.654273 1.937890 1.950068 0.5425766
#> [49,] 0.9595999 -0.7422829 1.887304 1.794829 2.128367 2.057179 0.5319272
#> [50,] 0.9570822 -0.9309629 2.076018 1.751524 1.767527 2.313386 0.5847070
#> [51,] 0.9709560 -0.6903834 1.957234 1.538792 2.024738 1.956733 0.5872545
#> [52,] 0.9920913 -0.8065503 1.589171 1.833705 2.007750 1.924565 0.6753279
#> [53,] 0.9693210 -0.9902179 2.000261 1.702651 2.048542 1.831388 0.4490391
#> [54,] 0.6745553 -0.8310121 2.131697 1.739619 1.871238 1.718252 0.5834831
#> [55,] 0.9113695 -0.9395162 2.229259 1.738378 2.345372 1.749114 0.5134592
#> [56,] 0.9004854 -0.4081808 2.240129 1.623039 2.166297 1.912650 0.5708480
#> [57,] 0.9728300 -0.6863845 1.785713 1.486197 1.769605 1.653543 0.6199279
#> [58,] 0.9743816 -0.8598771 1.574299 2.312789 1.654528 1.976010 0.7406975
#> [59,] 0.9341561 -0.5604109 2.023679 1.591369 1.961541 2.014248 0.5602467
#> [60,] 0.9349775 -0.9895314 1.845949 1.955237 2.107871 2.001888 0.4686762
#> [61,] 0.9560495 -0.8061916 1.756264 1.579666 2.074535 2.177315 0.4886237
#> [62,] 0.9945459 -0.8266822 1.694064 1.253005 2.026859 1.998841 0.5034967
#> [63,] 0.8652049 -0.8419817 1.746873 1.511866 1.781692 2.246844 0.6254742
#> [64,] 0.9625474 -0.7204887 2.021321 1.695890 2.316256 1.910311 0.4929317
#> [65,] 0.9883273 -0.9307958 2.082641 1.720413 2.056058 1.967253 0.4438644
#> [66,] 0.8017913 -0.9966729 2.196714 1.830430 2.069166 1.919774 0.5937950
#> [67,] 0.9357862 -0.7352854 2.251841 1.682189 1.980103 2.130198 0.5550117
#> [68,] 0.9871053 -0.8559998 1.965188 1.752562 2.072212 2.135551 0.3182215
#> [69,] 0.9482782 -0.8015607 2.110857 1.960294 2.457877 1.946282 0.5040165
#> [70,] 0.8122000 -0.7986266 1.936561 2.028310 2.034845 1.999985 0.5603122
#> [71,] 0.9344215 -0.7654327 1.882218 1.899149 2.319807 1.874472 0.4732485
#> [72,] 0.9373724 -0.9425771 1.978363 1.611343 2.267762 1.909805 0.5173838
#> [73,] 0.9478120 -0.9204906 2.311762 1.762212 2.325603 1.927763 0.5550627
#> [74,] 0.9558916 -0.8734623 1.942872 1.735849 2.055300 1.669529 0.5035401
#> [75,] 0.9409650 -0.6939841 1.757774 1.804823 2.066388 1.733775 0.5555570
#> [76,] 0.8335251 -0.8130979 2.006306 1.724000 1.802652 1.786837 0.5434985
#> [77,] 0.9650136 -0.8415538 1.697248 1.556270 1.850504 1.937444 0.5821920
#> [78,] 0.9109875 -0.8944437 1.977593 1.731018 1.989219 1.921683 0.5090669
#> [79,] 0.9823584 -0.9075523 2.106359 1.261726 2.074153 1.981471 0.5569617
#> [80,] 0.9898790 -0.8060299 1.930881 1.745941 2.047375 2.260124 0.6149701
#>            [,8]
#>  [1,] 0.5891540
#>  [2,] 0.5932398
#>  [3,] 0.5740067
#>  [4,] 0.4932558
#>  [5,] 0.5724593
#>  [6,] 0.5111063
#>  [7,] 0.6070918
#>  [8,] 0.3835398
#>  [9,] 0.6059201
#> [10,] 0.5227965
#> [11,] 0.4559322
#> [12,] 0.4472026
#> [13,] 0.5075802
#> [14,] 0.5657862
#> [15,] 0.4922075
#> [16,] 0.6136237
#> [17,] 0.5422139
#> [18,] 0.6781248
#> [19,] 0.4382591
#> [20,] 0.4807116
#> [21,] 0.6083456
#> [22,] 0.4345695
#> [23,] 0.5053175
#> [24,] 0.4703010
#> [25,] 0.4759132
#> [26,] 0.4771939
#> [27,] 0.4714255
#> [28,] 0.4841850
#> [29,] 0.5373168
#> [30,] 0.5610606
#> [31,] 0.5293563
#> [32,] 0.4777637
#> [33,] 0.4697557
#> [34,] 0.5999356
#> [35,] 0.6292046
#> [36,] 0.5395574
#> [37,] 0.5941473
#> [38,] 0.4484701
#> [39,] 0.6004296
#> [40,] 0.5099504
#> [41,] 0.5129706
#> [42,] 0.5624490
#> [43,] 0.4490732
#> [44,] 0.5941145
#> [45,] 0.4145766
#> [46,] 0.5647994
#> [47,] 0.6225884
#> [48,] 0.5452821
#> [49,] 0.5550127
#> [50,] 0.5502531
#> [51,] 0.5706835
#> [52,] 0.5624452
#> [53,] 0.5540878
#> [54,] 0.5467070
#> [55,] 0.6435912
#> [56,] 0.5764253
#> [57,] 0.3642791
#> [58,] 0.4200926
#> [59,] 0.4950916
#> [60,] 0.4825734
#> [61,] 0.4900090
#> [62,] 0.5530962
#> [63,] 0.4624670
#> [64,] 0.5413133
#> [65,] 0.5572864
#> [66,] 0.4316629
#> [67,] 0.4503390
#> [68,] 0.4720518
#> [69,] 0.3539278
#> [70,] 0.4661937
#> [71,] 0.5430519
#> [72,] 0.5245934
#> [73,] 0.5141972
#> [74,] 0.6446680
#> [75,] 0.3235886
#> [76,] 0.6071507
#> [77,] 0.5413652
#> [78,] 0.5936803
#> [79,] 0.5103008
#> [80,] 0.4287046
#> 
#> $member$storepop
#> list()
#> 
#> 
#> attr(,"class")
#> [1] "DEoptim"
#> 
#> $objective
#> [1] 251.8245
#> 
#> $residuals
#>                [,1]        [,2]
#>   [1,] -0.388166666 -3.28900833
#>   [2,] -0.294941565 -0.78805713
#>   [3,]  0.639488728  0.13594981
#>   [4,]  1.657267268  1.06889013
#>   [5,]  1.104341832 -0.86537238
#>   [6,]  1.348249058 -0.38552109
#>   [7,] -0.373241280 -1.31047071
#>   [8,]  0.037019346 -1.37388151
#>   [9,] -0.432385351 -0.87346595
#>  [10,] -0.903018779 -0.14284974
#>  [11,]  0.531660273 -1.09342967
#>  [12,] -0.691474640  0.31377427
#>  [13,] -1.029610100 -1.00635999
#>  [14,]  0.079301494  1.88008302
#>  [15,]  0.853918432  1.35253051
#>  [16,]  0.947176206 -1.11530065
#>  [17,] -1.086486287 -1.65311495
#>  [18,] -1.666144267 -0.07519907
#>  [19,] -1.525356480 -1.29748649
#>  [20,]  0.816920401 -1.01949344
#>  [21,]  0.187569626 -0.22653084
#>  [22,]  0.457328254  1.72503591
#>  [23,]  2.866760141  2.06054127
#>  [24,]  0.322287755 -0.86843708
#>  [25,]  2.370759224  0.64550907
#>  [26,]  0.708431113  1.15764002
#>  [27,] -0.705346637  2.24248731
#>  [28,]  1.987671741  0.58245936
#>  [29,]  0.011553975  0.47158316
#>  [30,]  0.825681346 -1.07848097
#>  [31,]  0.314791550  0.90048705
#>  [32,] -0.764262194 -1.42618422
#>  [33,] -1.535288006 -1.29977318
#>  [34,]  0.935741865 -1.12603779
#>  [35,]  0.706292992  0.17673491
#>  [36,] -0.280207358 -0.69128480
#>  [37,]  0.262494506  0.34330385
#>  [38,] -0.523671049 -0.42063382
#>  [39,]  0.126042542  1.41626756
#>  [40,]  3.814646158 -0.49344492
#>  [41,] -0.269796944  2.71449978
#>  [42,]  1.312854606  0.07328117
#>  [43,]  0.351705580  0.45932724
#>  [44,] -1.456229186  1.00208214
#>  [45,]  0.810755363  0.81382004
#>  [46,]  0.441217575 -0.77985325
#>  [47,]  0.126168331 -1.23337008
#>  [48,]  1.135707674  1.07777576
#>  [49,] -1.631373639 -0.01675685
#>  [50,] -0.701006378 -1.38826952
#>  [51,] -0.991299114 -1.90242953
#>  [52,]  2.177012179  2.50378068
#>  [53,]  0.506727120 -1.72689673
#>  [54,] -0.455802056 -0.01168924
#>  [55,]  0.016498014  0.28987105
#>  [56,] -0.872890385 -1.12020478
#>  [57,]  0.109312647  0.23482681
#>  [58,] -0.754309434 -0.76293657
#>  [59,]  1.240197680  1.35128632
#>  [60,]  0.560967499  0.83064325
#>  [61,]  1.384419828  1.56007094
#>  [62,]  0.246697079 -0.80044514
#>  [63,] -0.199061336  0.23097991
#>  [64,]  0.264424451 -0.49974501
#>  [65,] -2.053719987 -1.68065361
#>  [66,] -1.354730911  0.06530440
#>  [67,] -0.157260286  0.63950063
#>  [68,]  1.422880769  0.14088532
#>  [69,]  0.229895268 -0.53019953
#>  [70,] -0.375600466 -1.04388815
#>  [71,] -0.198108494 -0.19332820
#>  [72,]  3.499712918 -0.72987668
#>  [73,] -0.187772938  0.40229772
#>  [74,]  1.223088664 -0.40202020
#>  [75,]  0.795729581  1.98284944
#>  [76,] -0.691740097 -1.49712828
#>  [77,]  0.851702454 -1.38134553
#>  [78,]  0.059648223  0.37489347
#>  [79,]  1.372981010 -0.02483180
#>  [80,]  1.486966598 -0.07834847
#>  [81,]  1.187712856  0.86208181
#>  [82,]  0.953257941  2.85755040
#>  [83,] -0.359809439 -0.40064811
#>  [84,]  0.086706367  0.71746357
#>  [85,] -2.101752288 -0.95539229
#>  [86,] -0.298337787  0.07121661
#>  [87,]  0.202742449  0.43849041
#>  [88,] -1.343893125  0.95629435
#>  [89,]  0.132621408 -0.08994869
#>  [90,]  1.719751836 -0.99313523
#>  [91,]  0.003204632  0.64172317
#>  [92,]  1.387406091 -0.20457848
#>  [93,]  0.811919382 -0.87680551
#>  [94,] -0.508857833 -1.42212423
#>  [95,] -0.604150342 -0.24904403
#>  [96,]  0.806339585 -0.57731005
#>  [97,] -0.821038467 -1.16946375
#>  [98,] -0.662726735 -0.80897585
#>  [99,]  0.333661545  0.71865528
#> [100,]  1.715052011  0.73870060
#> 
#> $parameters
#>    alpha_1    alpha_2    beta_11    beta_21    beta_12    beta_22   gamma_11 
#>  0.9779260 -0.7421743  1.8943784  1.8058577  1.9425347  1.9304992  0.5034501 
#>   gamma_22   sigma_11   sigma_22 
#>  0.5053175  1.2151343  1.2652728 
#>