Fit a model to data

This method enables the user to fit a model (defined in model) on a particular dataset (defined in data). The estimation procedure makes use of numerical optimization using either DEoptim or nloptr, as specified by the user. Estimation proceeds through an optimization according to the output of the objective function of the provided model, as defined through objective_function, thus using least-squares as optimization standard.

Usage

fit(model, data, ...)

# S4 method for class 'model,dataset'
fit(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "DEoptim",
  ...
)

Arguments

model

Instance of the model-class, defining the model to evaluate the objective function for.

data

Instance of the dataset-class containing the data to fit the model to.

...

Arguments passed on to the control parameters of the optimizer, either to DEoptim.control or the opts argument of nloptr.

dynamics

Character denoting the structure of the dynamical matrices. Can either be "anisotropic" (completely free), "symmetric" (symmetric around the diagonal), and "isotropic" (diagonal). Note that this influences different parameters for different models, namely \(\Gamma\) for the exponential discounting model, \(N\) and \(K\) for the quasi-hyperbolic discounting model, and \(\Gamma\) and \(N\) for the double-exponential discounting model. Defaults to "isotropic".

covariance

Character denoting the structure of covariance matrix. Can either by "symmetric" (symmetric around the diagonal) and "isotropic" (diagonal). Defaults to "symmetric".

optimizer

Character denoting the optimizer to use for the estimation. Can either be "DEoptim" for the differential evolution algorithm in DEoptim or "nloptr" for the library implemented in nloptr. Defaults to "DEoptim".

Value

An named list containing an instance of the model-class with the estimated parameters ("model"), the results of the optimization procedure ("fit"), the value of the objective after ending the optimization procedure ("objective"), the residuals of the model ("residuals"), and a named vector containing the values of the estimated parameters linked to a character vector explaining their content ("parameters").

Details

Note that currently, least-squares estimation is assumed, meaning that the covariance matrix is left out of the objective function. If maximum-likelihood were needed instead, then this function would need to change.

Examples

# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  itermax = 50,
  trace = FALSE
)
#> $model
#> Model of class "exponential":
#> 
#> Dimension: 2
#> Number of predictors: 2
#> Number of parameters: 10
#> 
#> Parameters:
#>   alpha: |  0.9872914  |
#>          |  -0.7688216  |
#> 
#>   beta: | 1.93459  1.909332 |
#>         | 1.77778  2.053655 |
#> 
#>   gamma: | 0.5777155  0.00 |
#>          | 0.00  0.5264642 |
#> 
#> 
#> Covariance: | 1.228895  0.00 |
#>             | 0.00  1.228883 |
#> 
#> $fit
#> $optim
#> $optim$bestmem
#>       par1       par2       par3       par4       par5       par6       par7 
#>  0.9872914 -0.7688216  1.9345901  1.7777799  1.9093319  2.0536552  0.5777155 
#>       par8 
#>  0.5264642 
#> 
#> $optim$bestval
#> [1] 245.897
#> 
#> $optim$nfeval
#> [1] 4080
#> 
#> $optim$iter
#> [1] 50
#> 
#> 
#> $member
#> $member$lower
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>   -1   -1   -5   -5   -5   -5    0    0 
#> 
#> $member$upper
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>    1    1    5    5    5    5    1    1 
#> 
#> $member$bestmemit
#>          par1        par2     par3      par4     par5     par6      par7
#> 1  -0.5598897 -0.02447577 3.491405 1.1746574 1.977838 3.487849 0.6387811
#> 2   0.4136919  0.11228591 2.020463 0.3258226 2.319539 3.897326 0.5760894
#> 3   0.4136919  0.11228591 2.020463 0.3258226 2.319539 3.897326 0.5760894
#> 4   0.4831714 -0.90885788 3.713583 2.6202437 1.758339 3.184067 0.1470606
#> 5   0.5973556  0.17911982 1.519546 1.3119615 2.720025 2.075808 0.5417933
#> 6   0.5973556  0.17911982 1.519546 1.3119615 2.720025 2.075808 0.5417933
#> 7   0.5973556  0.17911982 1.519546 1.3119615 2.720025 2.075808 0.5417933
#> 8   0.5973556  0.17911982 1.519546 1.3119615 2.720025 2.075808 0.5417933
#> 9   0.9834479 -0.23886147 1.910067 1.0431339 1.316696 2.800730 0.7122438
#> 10  0.9834479 -0.23886147 1.910067 1.0431339 1.316696 2.800730 0.7122438
#> 11  0.4831714 -0.90885788 1.449349 2.6202437 1.758339 2.351578 0.3962310
#> 12  0.3518882 -0.22441536 2.289732 1.8426199 1.817646 2.501500 0.6967889
#> 13  0.4797392 -0.97076714 2.228514 1.3809625 2.003391 1.720969 0.5177864
#> 14  0.4797392 -0.97076714 2.228514 1.3809625 2.003391 1.720969 0.5177864
#> 15  0.8598577 -0.21802494 2.137977 2.0729057 2.041916 1.653759 0.6027381
#> 16  0.8598577 -0.21802494 2.137977 2.0729057 2.041916 1.653759 0.6027381
#> 17  0.8598577 -0.21802494 2.137977 2.0729057 2.041916 1.653759 0.6027381
#> 18  0.8598577 -0.21802494 2.137977 2.0729057 2.041916 1.653759 0.6027381
#> 19  0.4831714 -0.90885788 1.449349 2.0156781 2.072922 2.351578 0.6077158
#> 20  0.4831714 -0.90885788 1.449349 2.0156781 2.072922 2.351578 0.6077158
#> 21  0.8598577 -0.21802494 2.137977 2.0729057 2.041916 1.653759 0.4986214
#> 22  0.6107467 -0.89276041 2.137977 2.0729057 2.041916 1.653759 0.4986214
#> 23  0.6107467 -0.89276041 2.137977 2.0729057 2.041916 1.653759 0.4986214
#> 24  0.9778390 -0.54712376 1.887072 1.5379067 2.306163 2.066276 0.3281495
#> 25  0.9778390 -0.54712376 1.887072 1.5379067 2.306163 2.066276 0.3281495
#> 26  0.9778390 -0.54712376 1.887072 1.5379067 2.306163 2.066276 0.3281495
#> 27  0.8344329 -0.90885788 1.449349 2.0156781 2.072922 2.351578 0.6077158
#> 28  0.7800919 -0.68295998 1.886768 2.1807314 2.212228 2.053655 0.5777155
#> 29  0.7839848 -0.89276041 2.137977 2.0729057 2.041916 2.336855 0.5584885
#> 30  0.7839848 -0.89276041 2.137977 2.0729057 2.041916 2.336855 0.5584885
#> 31  0.8918779 -0.58130952 1.737438 1.8645367 1.764439 2.169984 0.5231718
#> 32  0.8918779 -0.58130952 1.737438 1.8645367 1.764439 2.169984 0.5231718
#> 33  0.8918779 -0.58130952 1.737438 1.8645367 1.764439 2.169984 0.5231718
#> 34  0.9940450 -0.67319965 1.980922 1.4347724 2.050435 2.099729 0.5438170
#> 35  0.9940450 -0.67319965 1.980922 1.4347724 2.050435 2.099729 0.5438170
#> 36  0.9940450 -0.67319965 1.980922 1.4347724 2.050435 2.099729 0.5438170
#> 37  0.9940450 -0.67319965 1.980922 1.4347724 2.050435 2.099729 0.5438170
#> 38  0.8171176 -0.56691101 1.729745 1.8701208 1.985969 1.910595 0.5917083
#> 39  0.8918779 -0.58130952 1.737438 1.8645367 2.078228 2.020212 0.5231718
#> 40  0.8171176 -0.76096646 1.729745 1.8701208 1.985969 1.910595 0.5917083
#> 41  0.8171176 -0.76096646 1.729745 1.8701208 1.985969 1.910595 0.5917083
#> 42  0.8171176 -0.76096646 1.729745 1.8701208 1.985969 1.910595 0.5917083
#> 43  0.9778390 -0.97136117 1.892063 1.9455528 2.109873 2.202724 0.5625089
#> 44  0.9342405 -0.75652787 1.904847 1.9154347 1.926812 2.111149 0.5926273
#> 45  0.9342405 -0.75652787 1.896105 1.9154347 1.926812 2.111149 0.5926273
#> 46  0.9342405 -0.75652787 1.896105 1.9154347 1.926812 2.111149 0.5493330
#> 47  0.9342405 -0.75652787 1.896105 1.9154347 1.926812 2.111149 0.5493330
#> 48  0.9747591 -0.72773418 1.928217 1.8812199 2.003391 2.003302 0.5177864
#> 49  0.9747591 -0.72773418 1.928217 1.8812199 2.003391 2.003302 0.5177864
#> 50  0.9747591 -0.72773418 1.928217 1.8812199 2.003391 2.003302 0.5177864
#>         par8
#> 1  0.5693208
#> 2  0.2289416
#> 3  0.2236577
#> 4  0.4549908
#> 5  0.1740448
#> 6  0.1740448
#> 7  0.1740448
#> 8  0.1740448
#> 9  0.2126528
#> 10 0.2126528
#> 11 0.4549908
#> 12 0.3943770
#> 13 0.3434863
#> 14 0.3434863
#> 15 0.4295232
#> 16 0.4295232
#> 17 0.4295232
#> 18 0.4295232
#> 19 0.4549908
#> 20 0.4549908
#> 21 0.4295232
#> 22 0.4295232
#> 23 0.4295232
#> 24 0.5077631
#> 25 0.5077631
#> 26 0.5077631
#> 27 0.4549908
#> 28 0.3806096
#> 29 0.4295232
#> 30 0.4295232
#> 31 0.5355225
#> 32 0.5355225
#> 33 0.5355225
#> 34 0.4959109
#> 35 0.4959109
#> 36 0.4959109
#> 37 0.4959109
#> 38 0.5192731
#> 39 0.5355225
#> 40 0.5192731
#> 41 0.5192731
#> 42 0.5192731
#> 43 0.5077631
#> 44 0.4950096
#> 45 0.4950096
#> 46 0.4950096
#> 47 0.4950096
#> 48 0.5273298
#> 49 0.5273298
#> 50 0.5273298
#> 
#> $member$bestvalit
#>  [1] 1257.3762  925.9520  925.0972  863.4333  514.2897  514.2897  514.2897
#>  [8]  514.2897  486.1085  486.1085  411.5946  397.8031  363.9765  363.9765
#> [15]  321.9731  321.9731  321.9731  321.9731  321.7360  321.7360  321.0218
#> [22]  315.5696  315.5696  298.9019  298.9019  298.9019  291.3252  287.3736
#> [29]  277.4772  277.4772  272.0766  272.0766  272.0766  265.4249  265.4249
#> [36]  265.4249  265.4249  263.3548  261.1350  256.3333  256.3333  256.3333
#> [43]  252.6750  249.5221  249.4589  249.2190  249.2190  247.3589  247.3589
#> [50]  247.3589
#> 
#> $member$pop
#>             [,1]       [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.86215520 -0.5649599 1.895390 1.877525 2.042633 1.807983 0.5381185
#>  [2,] 0.88622191 -0.8537091 2.043371 1.919369 1.997360 2.155297 0.5570726
#>  [3,] 0.98729141 -0.7688216 1.934590 1.777780 1.909332 2.053655 0.5777155
#>  [4,] 0.93424050 -0.7565279 1.896105 1.915435 1.926812 2.111149 0.5493330
#>  [5,] 0.76452769 -0.7704259 1.893910 1.667531 1.949789 2.046793 0.5504899
#>  [6,] 0.90097650 -0.7822716 1.856280 1.878642 2.077703 1.908644 0.4654347
#>  [7,] 0.77392132 -0.4144557 1.986912 1.883268 1.788737 2.039081 0.6059115
#>  [8,] 0.78398475 -0.8927604 2.137977 2.072906 2.011540 1.984077 0.5120317
#>  [9,] 0.95683078 -0.9580746 1.925349 2.063250 1.951269 2.367785 0.5523580
#> [10,] 0.80073274 -0.9299455 1.881365 2.019027 1.999658 1.992104 0.5769670
#> [11,] 0.78041345 -0.8205746 1.945949 2.046871 1.972476 1.942486 0.6089790
#> [12,] 0.94091379 -0.8361021 2.002147 1.849518 1.789482 1.979271 0.4502013
#> [13,] 0.96360251 -0.6471620 1.958066 1.705904 1.831274 2.196274 0.5886506
#> [14,] 0.92513955 -0.9338389 1.962980 2.119541 2.005400 1.884017 0.4827553
#> [15,] 0.88946764 -0.8310833 1.865788 1.890921 2.001757 2.015560 0.5786409
#> [16,] 0.96358098 -0.9550727 1.877691 1.954072 2.083963 2.183117 0.3829999
#> [17,] 0.94615336 -0.5395079 1.678320 2.162199 2.138050 2.151165 0.5859256
#> [18,] 0.87083247 -0.9505725 1.714758 1.768848 1.797283 2.045270 0.6107436
#> [19,] 0.85193708 -0.8634298 1.682979 1.817244 1.975865 2.101652 0.5689979
#> [20,] 0.94607821 -0.7454659 1.927875 2.040851 2.194364 1.956907 0.5756121
#> [21,] 0.89187789 -0.7103658 1.876121 2.079903 1.987470 2.020212 0.5417105
#> [22,] 0.82873847 -0.7583142 1.650531 2.125105 2.036585 2.082888 0.5893989
#> [23,] 0.87241777 -0.6654646 2.013433 1.874129 1.953699 2.151815 0.5787949
#> [24,] 0.97248708 -0.6861545 2.089150 1.692142 2.208534 1.928338 0.4381566
#> [25,] 0.97205171 -0.6525721 1.979931 1.825462 2.088726 1.818825 0.5054397
#> [26,] 0.79874613 -0.7826818 1.883318 2.053052 2.087856 1.990744 0.6315470
#> [27,] 0.98886697 -0.7774524 2.224308 1.934433 2.090302 2.086457 0.5873918
#> [28,] 0.94284223 -0.8141761 1.918593 2.084367 2.023522 2.138592 0.5112740
#> [29,] 0.83067482 -0.9193881 1.366601 1.712637 1.505249 2.016713 0.6304776
#> [30,] 0.88253359 -0.8020048 1.864412 1.912321 1.858617 2.186542 0.6386517
#> [31,] 0.98403302 -0.7909325 1.813789 2.172636 2.115808 2.235099 0.5832024
#> [32,] 0.05058777 -0.9026528 1.768414 1.616469 1.835432 2.093612 0.7075852
#> [33,] 0.99046335 -0.9490385 1.880919 1.785130 1.995567 2.197378 0.5131316
#> [34,] 0.90497933 -0.9729992 2.125531 1.917748 2.227085 2.201514 0.4987854
#> [35,] 0.85926023 -0.8263796 2.231311 1.716192 1.891889 1.844685 0.5340511
#> [36,] 0.92833830 -0.9177040 1.855582 1.702881 1.809711 2.026229 0.5597502
#> [37,] 0.98761045 -0.5780722 1.848419 2.139070 1.931007 2.068543 0.6326484
#> [38,] 0.76689614 -0.1735860 1.940761 1.729431 2.053312 1.982197 0.5576495
#> [39,] 0.97783903 -0.9713612 1.892063 1.937211 2.109873 2.202724 0.5625089
#> [40,] 0.90907180 -0.7475765 2.025011 2.067950 1.934028 2.082598 0.6015952
#> [41,] 0.90706326 -0.9070002 1.672835 1.833150 2.179028 2.001889 0.4862732
#> [42,] 0.89859489 -0.7305492 1.808387 1.494663 1.995453 1.893258 0.5556546
#> [43,] 0.80791854 -0.8029383 2.060438 1.666357 1.921650 2.057198 0.6071977
#> [44,] 0.94853942 -0.8274373 2.081600 1.819983 1.791165 2.294230 0.5719043
#> [45,] 0.92099836 -0.9310654 1.698342 1.751414 1.917607 1.954812 0.6547649
#> [46,] 0.99144733 -0.8861993 1.803341 2.134097 2.166707 1.890758 0.5125648
#> [47,] 0.84323681 -0.9182311 2.089114 2.099133 1.892522 2.114060 0.5593802
#> [48,] 0.92694411 -0.9412540 1.991055 1.714963 2.104946 1.948993 0.5351253
#> [49,] 0.74537681 -0.7737361 2.112777 1.863842 1.976154 2.022893 0.5847316
#> [50,] 0.90294122 -0.8789047 1.865183 1.890552 2.010987 1.959681 0.5264863
#> [51,] 0.99976369 -0.5020192 1.938427 1.767092 2.225043 1.873591 0.4446457
#> [52,] 0.93545923 -0.7111467 1.986151 1.908373 1.889246 2.215229 0.5671522
#> [53,] 0.92270829 -0.8376773 1.908692 1.965495 1.825101 2.132001 0.5889230
#> [54,] 0.87067169 -0.6992242 1.495016 1.819030 1.512264 2.413091 0.6163759
#> [55,] 0.80645387 -0.9996252 1.867802 1.919121 2.205966 1.936457 0.5430103
#> [56,] 0.89620069 -0.6453770 2.030223 1.875684 1.981572 1.992677 0.5546188
#> [57,] 0.93423129 -0.6873326 1.861837 1.817433 2.030779 2.129641 0.5671579
#> [58,] 0.99404498 -0.6731997 1.980922 1.940430 1.877890 1.953971 0.5350555
#> [59,] 0.81061140 -0.8111950 1.995386 1.808768 2.070758 2.109815 0.4882570
#> [60,] 0.74104966 -0.8564981 1.880260 1.881335 2.051530 2.153137 0.5867233
#> [61,] 0.98545463 -0.8130370 2.050864 2.028604 2.007158 2.008869 0.5194098
#> [62,] 0.81323439 -0.7418742 1.884155 1.844858 1.984671 2.069070 0.5693628
#> [63,] 0.97842533 -0.9459842 1.845777 1.821544 1.567311 2.048002 0.5636338
#> [64,] 0.97475912 -0.7277342 1.928217 1.881220 2.003391 2.003302 0.5177864
#> [65,] 0.81773154 -0.7972171 1.974321 1.824191 1.866960 2.125548 0.5577642
#> [66,] 0.95454462 -0.7479788 1.674147 1.800496 1.985969 1.910595 0.5917083
#> [67,] 0.84652335 -0.7714544 1.774586 1.724367 2.009215 1.914979 0.5000129
#> [68,] 0.91348767 -0.6031601 1.567620 1.639052 1.827025 2.040953 0.6075789
#> [69,] 0.86423844 -0.8388433 1.783505 2.126089 2.065287 2.125205 0.5734262
#> [70,] 0.90200488 -0.9061056 1.943153 2.001913 2.087775 2.005302 0.5617929
#> [71,] 0.97628800 -0.8476841 1.919599 1.632774 1.966519 1.816411 0.5348278
#> [72,] 0.91378689 -0.6274337 1.570245 1.988035 1.849809 2.138715 0.5910961
#> [73,] 0.90159339 -0.8892837 2.046416 1.761520 2.100087 1.832006 0.4103598
#> [74,] 0.99799183 -0.8350573 1.834102 1.724461 2.103661 2.054227 0.5220083
#> [75,] 0.92174003 -0.7965370 1.957296 1.728806 1.985635 2.001301 0.5167244
#> [76,] 0.99620944 -0.5573512 2.018543 1.825546 2.005491 1.923407 0.5175180
#> [77,] 0.75165245 -0.7871098 1.843063 1.445907 2.266356 1.921170 0.5377896
#> [78,] 0.84855645 -0.6456596 1.941332 1.844220 2.008859 2.123041 0.5407947
#> [79,] 0.90171685 -0.7951512 1.772627 2.100828 2.347630 1.684501 0.4852474
#> [80,] 0.71037286 -0.7447906 1.879761 2.095985 1.716020 1.850476 0.5806001
#>            [,8]
#>  [1,] 0.4758175
#>  [2,] 0.3737742
#>  [3,] 0.5264642
#>  [4,] 0.4950096
#>  [5,] 0.5730924
#>  [6,] 0.5256580
#>  [7,] 0.4533841
#>  [8,] 0.5768800
#>  [9,] 0.5207940
#> [10,] 0.5666487
#> [11,] 0.4752181
#> [12,] 0.5109559
#> [13,] 0.4478485
#> [14,] 0.5891873
#> [15,] 0.4771468
#> [16,] 0.4991344
#> [17,] 0.4737417
#> [18,] 0.4951139
#> [19,] 0.4256158
#> [20,] 0.5331476
#> [21,] 0.5355225
#> [22,] 0.5168241
#> [23,] 0.4891270
#> [24,] 0.5522210
#> [25,] 0.5579899
#> [26,] 0.5238212
#> [27,] 0.5217052
#> [28,] 0.5178894
#> [29,] 0.6373011
#> [30,] 0.4741509
#> [31,] 0.4188089
#> [32,] 0.4952132
#> [33,] 0.5425274
#> [34,] 0.5041233
#> [35,] 0.6463777
#> [36,] 0.6388133
#> [37,] 0.5089124
#> [38,] 0.5267043
#> [39,] 0.5077631
#> [40,] 0.5008245
#> [41,] 0.6155852
#> [42,] 0.5523729
#> [43,] 0.4690734
#> [44,] 0.4983635
#> [45,] 0.5182006
#> [46,] 0.5542192
#> [47,] 0.5437212
#> [48,] 0.6574265
#> [49,] 0.4782794
#> [50,] 0.5753115
#> [51,] 0.4989345
#> [52,] 0.4836750
#> [53,] 0.4927333
#> [54,] 0.4776846
#> [55,] 0.6147463
#> [56,] 0.4865844
#> [57,] 0.5230287
#> [58,] 0.4959109
#> [59,] 0.4172193
#> [60,] 0.7062244
#> [61,] 0.5147164
#> [62,] 0.5787740
#> [63,] 0.5016236
#> [64,] 0.5273298
#> [65,] 0.5676649
#> [66,] 0.5192731
#> [67,] 0.5372650
#> [68,] 0.5254579
#> [69,] 0.4522788
#> [70,] 0.4558078
#> [71,] 0.6884785
#> [72,] 0.4427714
#> [73,] 0.5067151
#> [74,] 0.3599719
#> [75,] 0.4308704
#> [76,] 0.4697121
#> [77,] 0.6072956
#> [78,] 0.4347974
#> [79,] 0.5502426
#> [80,] 0.4458565
#> 
#> $member$storepop
#> list()
#> 
#> 
#> attr(,"class")
#> [1] "DEoptim"
#> 
#> $objective
#> [1] 245.897
#> 
#> $residuals
#>                 [,1]         [,2]
#>   [1,] -0.1786176653 -0.703601547
#>   [2,]  0.5560419515 -0.004199962
#>   [3,]  1.4399935083  0.944777147
#>   [4,]  1.0198432072 -0.992842484
#>   [5,]  1.5000888453 -0.508053479
#>   [6,] -0.1580727243 -1.476037247
#>   [7,]  0.1723295754 -1.232536972
#>   [8,] -0.2351816830 -0.901211389
#>   [9,] -1.0146586204 -0.411212944
#>  [10,]  0.0125671303 -1.329411870
#>  [11,] -1.0493671503  0.222793107
#>  [12,] -1.2383434692 -1.123175517
#>  [13,] -0.0272702812  1.970297228
#>  [14,]  1.2058113470  1.482683137
#>  [15,]  1.2963271415 -0.866767457
#>  [16,] -0.7277729213 -1.715245554
#>  [17,] -1.5727199252 -0.002172433
#>  [18,] -1.5031299386 -1.403475647
#>  [19,]  0.6730141568 -1.162716882
#>  [20,] -0.0001042806 -0.531793213
#>  [21,]  0.3391005893  1.404343088
#>  [22,]  2.5487816098  1.769213861
#>  [23,] -0.3026397032 -1.113491135
#>  [24,]  1.8174295373  0.536229308
#>  [25,]  0.0714252298  1.248622040
#>  [26,] -0.9809526679  2.227231805
#>  [27,]  1.7835322926  0.610685493
#>  [28,]  0.0890298601  0.597601166
#>  [29,]  0.9767763333 -1.096618996
#>  [30,]  0.6092845063  0.902199134
#>  [31,] -0.7293836439 -1.416280880
#>  [32,] -1.6781323744 -1.239439422
#>  [33,]  0.8876259115 -1.193550353
#>  [34,]  0.7235016435  0.251982533
#>  [35,] -0.0524125562 -0.704591865
#>  [36,]  0.4664983665  0.091306588
#>  [37,] -0.8215657954 -0.672768409
#>  [38,] -0.2158355451  1.119107628
#>  [39,]  3.5456071529 -0.735372212
#>  [40,] -0.4320407914  2.490378440
#>  [41,]  1.1782032961  0.092400598
#>  [42,]  0.3723449153  0.396350888
#>  [43,] -1.7176973355  1.142636298
#>  [44,]  0.8363892209  0.985015174
#>  [45,]  0.7463732010 -0.555219485
#>  [46,]  0.5692548123 -1.021593608
#>  [47,]  1.5930272496  1.060525908
#>  [48,] -1.6420932033 -0.219850931
#>  [49,] -0.9782915035 -1.463259379
#>  [50,] -1.3318267509 -2.099239233
#>  [51,]  1.9057169068  2.297050075
#>  [52,]  0.1011794845 -1.878544855
#>  [53,] -0.7648423867 -0.058643045
#>  [54,] -0.3472639145  0.362915834
#>  [55,] -1.3159024176 -1.200050328
#>  [56,] -0.2966857247  0.037616573
#>  [57,] -1.4066393011 -0.806564747
#>  [58,]  0.5765311321  1.248425185
#>  [59,]  0.1121686065  0.815224491
#>  [60,]  1.1832734262  1.486701207
#>  [61,]  0.1246863102 -0.860046577
#>  [62,] -0.2767298311  0.290022209
#>  [63,]  0.2439290389 -0.533956541
#>  [64,] -2.0903241393 -1.539111782
#>  [65,] -1.2014882669  0.322748548
#>  [66,] -0.1246894725  1.013632648
#>  [67,]  1.4913486152  0.584017602
#>  [68,]  0.0720879715 -0.272636532
#>  [69,] -0.6443615955 -0.892340283
#>  [70,] -0.4323272697  0.189933504
#>  [71,]  3.6186236729 -0.603363129
#>  [72,] -0.3336569803  0.336941257
#>  [73,]  0.7442139115 -0.630584748
#>  [74,]  0.3263578694  1.766449170
#>  [75,] -1.1100421134 -1.711764226
#>  [76,]  0.5854422226 -1.496520617
#>  [77,] -0.4142501941  0.053715673
#>  [78,]  0.8649757842 -0.321568649
#>  [79,]  1.0171409438 -0.254400577
#>  [80,]  0.8546610668  1.059157335
#>  [81,]  1.3215099057  2.987485149
#>  [82,] -0.1496927570 -0.209811971
#>  [83,]  0.3386027180  0.878972260
#>  [84,] -1.8411319546 -1.099526436
#>  [85,] -0.1937074278 -0.066620453
#>  [86,]  0.2443493930  0.312500152
#>  [87,] -1.4677195002  0.953164466
#>  [88,]  0.1249700070 -0.264214351
#>  [89,]  1.5593683135 -1.084370316
#>  [90,] -0.0208808881  0.657289420
#>  [91,]  1.4625513611 -0.098482121
#>  [92,]  0.9014347511 -0.874396803
#>  [93,] -0.2564048982 -1.276480084
#>  [94,] -0.4770330655  0.058087634
#>  [95,]  0.8877996613 -0.498291428
#>  [96,] -0.9996002671 -1.249340183
#>  [97,] -1.0061677647 -0.715819743
#>  [98,]  0.2578726912  0.844473496
#>  [99,]  1.7827966581  1.108280823
#> [100,] -1.0235193364 -2.304044628
#> 
#> $parameters
#>    alpha_1    alpha_2    beta_11    beta_21    beta_12    beta_22   gamma_11 
#>  0.9872914 -0.7688216  1.9345901  1.7777799  1.9093319  2.0536552  0.5777155 
#>   gamma_22   sigma_11   sigma_22 
#>  0.5264642  1.2288951  1.2288827 
#>