Fit a model to data

This method enables the user to fit a model (defined in model) on a particular dataset (defined in data). The estimation procedure makes use of numerical optimization using either DEoptim or nloptr, as specified by the user. Estimation proceeds through an optimization according to the output of the objective function of the provided model, as defined through objective_function, thus using least-squares as optimization standard.

Usage

fit(model, data, ...)

# S4 method for class 'model,dataset'
fit(
  model,
  data,
  dynamics = "isotropic",
  covariance = "symmetric",
  optimizer = "DEoptim",
  ...
)

Arguments

model

Instance of the model-class, defining the model to evaluate the objective function for.

data

Instance of the dataset-class containing the data to fit the model to.

...

Arguments passed on to the control parameters of the optimizer, either to DEoptim.control or the opts argument of nloptr.

dynamics

Character denoting the structure of the dynamical matrices. Can either be "anisotropic" (completely free), "symmetric" (symmetric around the diagonal), and "isotropic" (diagonal). Note that this influences different parameters for different models, namely \(\Gamma\) for the exponential discounting model, \(N\) and \(K\) for the quasi-hyperbolic discounting model, and \(\Gamma\) and \(N\) for the double-exponential discounting model. Defaults to "isotropic".

covariance

Character denoting the structure of covariance matrix. Can either by "symmetric" (symmetric around the diagonal) and "isotropic" (diagonal). Defaults to "symmetric".

optimizer

Character denoting the optimizer to use for the estimation. Can either be "DEoptim" for the differential evolution algorithm in DEoptim or "nloptr" for the library implemented in nloptr. Defaults to "DEoptim".

Value

An named list containing an instance of the model-class with the estimated parameters ("model"), the results of the optimization procedure ("fit"), the value of the objective after ending the optimization procedure ("objective"), the residuals of the model ("residuals"), and a named vector containing the values of the estimated parameters linked to a character vector explaining their content ("parameters").

Details

Note that currently, least-squares estimation is assumed, meaning that the covariance matrix is left out of the objective function. If maximum-likelihood were needed instead, then this function would need to change.

Examples

# Simulate data to use for this example
data <- simulate(
  quasi_hyperbolic(
    parameters = list(
      "alpha" = c(1, -1) ,
      "beta" = matrix(2, nrow = 2, ncol = 2),
      "nu" = diag(2) * 0.75,
      "kappa" = diag(2) * 0.5
    ),
    covariance = matrix(c(1, 0.25, 0.25, 1), nrow = 2, ncol = 2)
  ),
  X = matrix(rnorm(200), nrow = 100, ncol = 2)
)

# Evaluate the objective function for an exponential model with a particular
# set of parameters
fit(
  exponential(d = 2, k = 2),
  data,
  dynamics = "isotropic",
  covariance = "isotropic",
  itermax = 50,
  trace = FALSE
)
#> $model
#> Model of class "exponential":
#> 
#> Dimension: 2
#> Number of predictors: 2
#> Number of parameters: 10
#> 
#> Parameters:
#>   alpha: |  0.8968625  |
#>          |  -0.6060573  |
#> 
#>   beta: | 1.71363  1.879813 |
#>         | 1.955257  1.924812 |
#> 
#>   gamma: | 0.5684217  0.00 |
#>          | 0.00  0.5639565 |
#> 
#> 
#> Covariance: | 1.417284  0.00 |
#>             | 0.00  1.471387 |
#> 
#> $fit
#> $optim
#> $optim$bestmem
#>       par1       par2       par3       par4       par5       par6       par7 
#>  0.8968625 -0.6060573  1.7136295  1.9552574  1.8798134  1.9248118  0.5684217 
#>       par8 
#>  0.5639565 
#> 
#> $optim$bestval
#> [1] 292.3204
#> 
#> $optim$nfeval
#> [1] 4080
#> 
#> $optim$iter
#> [1] 50
#> 
#> 
#> $member
#> $member$lower
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>   -1   -1   -5   -5   -5   -5    0    0 
#> 
#> $member$upper
#> par1 par2 par3 par4 par5 par6 par7 par8 
#>    1    1    5    5    5    5    1    1 
#> 
#> $member$bestmemit
#>          par1         par2       par3      par4     par5      par6      par7
#> 1  -0.9417917 -0.097327391  2.5967796 2.2790522 3.784999 2.9104537 0.5611449
#> 2  -0.9417917 -0.097327391  2.5967796 2.2790522 3.615814 2.9104537 0.5611449
#> 3   0.5183419 -0.721281365  0.4542524 2.1908077 2.385946 4.0810477 0.1813332
#> 4   0.5183419 -0.721281365  0.4542524 2.1908077 2.385946 4.0810477 0.1813332
#> 5   0.5183419 -0.721281365  0.4542524 2.1908077 2.385946 4.0810477 0.1813332
#> 6   0.8558905  0.567455030 -0.2601943 2.3526032 2.311495 2.7506233 0.6532899
#> 7   0.5606726 -0.882812958  2.5967796 2.2790522 3.615814 0.8918614 0.5611449
#> 8   0.5606726 -0.882812958  2.5967796 2.2790522 3.615814 0.8918614 0.3822907
#> 9   0.5606726 -0.882812958  2.5967796 2.2790522 3.615814 0.8918614 0.3822907
#> 10  0.5606726 -0.882812958  2.5967796 2.2790522 3.615814 0.8918614 0.3822907
#> 11  0.7643729 -0.392891109  1.6587556 0.4207581 2.061686 2.7041784 0.4681866
#> 12  0.6307262  0.133173653  2.0951504 1.1771932 1.787160 1.7052229 0.1984241
#> 13  0.6307262  0.133173653  2.0951504 1.1771932 1.787160 1.7052229 0.1984241
#> 14  0.6554052  0.111866520  1.9639664 1.6544451 2.310906 1.4106775 0.6740426
#> 15  0.6554052  0.111866520  1.9639664 1.6544451 2.310906 1.4106775 0.6740426
#> 16  0.6554052  0.111866520  1.9639664 1.6544451 2.310906 1.4106775 0.6740426
#> 17  0.7399687  0.008267742  1.9708147 1.4978352 2.370901 2.1802219 0.6102171
#> 18  0.8870643 -0.786095969  1.4963376 1.9271983 2.353658 2.9977051 0.6153632
#> 19  0.6307262 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.5051050
#> 20  0.6307262 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 21  0.6307262 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 22  0.6307262 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 23  0.6307262 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 24  0.7343758 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 25  0.7343758 -0.100056476  2.0951504 1.9600949 1.706530 1.7052229 0.6302186
#> 26  0.8870231 -0.303939917  1.6095078 2.1908077 2.385946 1.7790780 0.5787997
#> 27  0.8870231 -0.303939917  1.6095078 2.1908077 2.234291 1.7790780 0.5787997
#> 28  0.8839170 -0.807808780  1.6024638 1.7323458 1.470831 1.7265554 0.6542074
#> 29  0.8839170 -0.807808780  1.6024638 1.7323458 1.470831 1.9783033 0.6542074
#> 30  0.9592531 -0.807808780  1.6024638 1.7323458 1.470831 1.9783033 0.6542074
#> 31  0.9592531 -0.807808780  1.6024638 1.7323458 1.470831 1.9783033 0.6542074
#> 32  0.9592531 -0.807808780  1.6024638 1.7323458 1.470831 1.9783033 0.6627059
#> 33  0.8747382 -0.323040394  1.7523887 2.0200760 1.603474 1.8837267 0.6421238
#> 34  0.8774739 -0.923718423  1.9415571 2.0204419 1.755232 2.1143565 0.6354730
#> 35  0.8774739 -0.923718423  1.9415571 2.0204419 1.755232 2.1143565 0.6354730
#> 36  0.8774739 -0.923718423  1.9415571 2.0204419 1.755232 2.1143565 0.6354730
#> 37  0.8774739 -0.923718423  1.9415571 2.0204419 1.755232 2.1143565 0.6354730
#> 38  0.8774739 -0.923718423  1.9415571 2.0204419 1.755232 2.1143565 0.6354730
#> 39  0.9537830 -0.394014639  1.6247045 2.0200760 1.603474 1.8837267 0.6421238
#> 40  0.9537830 -0.394014639  1.6247045 2.0200760 1.603474 1.8837267 0.6421238
#> 41  0.8451026 -0.507991261  1.8997738 1.9399931 1.756061 1.8754690 0.5912897
#> 42  0.9537830 -0.550857323  1.6247045 2.0200760 1.836402 1.8837267 0.6421238
#> 43  0.9537830 -0.550857323  1.6247045 2.0200760 1.836402 1.8837267 0.6421238
#> 44  0.9537830 -0.550857323  1.6247045 2.0200760 1.836402 1.8837267 0.6421238
#> 45  0.9537830 -0.550857323  1.6247045 2.0200760 1.836402 1.8837267 0.6421238
#> 46  0.9537830 -0.550857323  1.6247045 2.0200760 1.836402 1.8837267 0.6421238
#> 47  0.9382532 -0.795502271  1.7235027 2.0006141 1.982200 2.2450344 0.5669336
#> 48  0.8968625 -0.606057274  1.7136295 1.9552574 1.879813 1.9248118 0.5684217
#> 49  0.8968625 -0.606057274  1.7136295 1.9552574 1.879813 1.9248118 0.5684217
#> 50  0.8968625 -0.606057274  1.7136295 1.9552574 1.879813 1.9248118 0.5684217
#>          par8
#> 1  0.54244168
#> 2  0.54244168
#> 3  0.18670618
#> 4  0.18670618
#> 5  0.18670618
#> 6  0.01409418
#> 7  0.54244168
#> 8  0.54244168
#> 9  0.54244168
#> 10 0.54244168
#> 11 0.35497728
#> 12 0.47061544
#> 13 0.47061544
#> 14 0.49424428
#> 15 0.49424428
#> 16 0.49424428
#> 17 0.28596765
#> 18 0.38398322
#> 19 0.47061544
#> 20 0.47061544
#> 21 0.47061544
#> 22 0.47061544
#> 23 0.47061544
#> 24 0.47061544
#> 25 0.47061544
#> 26 0.56681074
#> 27 0.56681074
#> 28 0.61805752
#> 29 0.61805752
#> 30 0.61805752
#> 31 0.61805752
#> 32 0.58940563
#> 33 0.51203144
#> 34 0.52498465
#> 35 0.52498465
#> 36 0.52498465
#> 37 0.52498465
#> 38 0.52498465
#> 39 0.51203144
#> 40 0.51203144
#> 41 0.51620769
#> 42 0.51203144
#> 43 0.51203144
#> 44 0.51203144
#> 45 0.51203144
#> 46 0.51203144
#> 47 0.49078581
#> 48 0.56395652
#> 49 0.56395652
#> 50 0.56395652
#> 
#> $member$bestvalit
#>  [1] 1388.2834 1302.5724 1294.1578 1294.1578 1294.1578 1262.2014 1125.8631
#>  [8]  896.9869  896.9869  896.9869  770.4298  700.9462  700.9462  551.8621
#> [15]  551.8621  551.8621  513.1798  455.6611  400.3743  383.7451  383.7451
#> [22]  383.7451  383.7451  377.6998  377.6998  357.7244  338.0687  329.0264
#> [29]  324.7871  322.4210  322.4210  318.7923  317.2046  313.2904  313.2904
#> [36]  313.2904  313.2904  313.2904  311.6586  311.6586  305.9591  303.5694
#> [43]  303.5694  303.5694  303.5694  303.5694  297.4443  292.3204  292.3204
#> [50]  292.3204
#> 
#> $member$pop
#>            [,1]        [,2]      [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,] 0.9864012 -0.82303379 1.6353699 1.749344 2.011004 1.969382 0.5410128
#>  [2,] 0.9824749 -0.66638661 1.6420477 1.840792 1.759798 1.829131 0.5305889
#>  [3,] 0.9293720 -0.49651813 1.6570562 2.335452 1.810606 2.002275 0.6038358
#>  [4,] 0.9425738 -0.79117750 1.9119262 1.729175 2.195771 1.847267 0.4868114
#>  [5,] 0.9890515 -0.81793115 1.2781071 2.040326 1.661597 2.162981 0.7210854
#>  [6,] 0.7669857 -0.51816469 2.2133449 1.890309 2.875009 1.988306 0.2866941
#>  [7,] 0.7298099 -0.56146484 1.5976227 2.246816 1.764905 2.348810 0.6373780
#>  [8,] 0.8156650 -0.77426296 1.7287913 2.381144 1.896134 2.157401 0.6278177
#>  [9,] 0.7349235 -0.60586431 1.2330579 2.008577 1.777358 2.167455 0.7269024
#> [10,] 0.9178049 -0.96866732 1.8122334 2.048638 1.998057 2.005328 0.5782985
#> [11,] 0.9435678 -0.46923635 1.9335818 2.144456 1.528478 1.949951 0.6137972
#> [12,] 0.8921738 -0.58949145 1.5943190 1.839216 1.847465 2.041486 0.5720824
#> [13,] 0.2169441 -0.55234048 1.2777497 2.023347 1.441556 2.264694 0.6609915
#> [14,] 0.9187811 -0.80113704 1.7139530 2.068820 2.558354 2.079803 0.5077507
#> [15,] 0.9032575 -0.60207977 1.7473054 1.747860 1.430720 2.028365 0.5524614
#> [16,] 0.8353319 -0.43295316 2.0467460 1.829736 2.000596 2.326077 0.5027762
#> [17,] 0.7192535 -0.65185620 1.9501803 2.132016 2.084345 1.946884 0.5236696
#> [18,] 0.9718969 -0.36450255 1.8134884 1.923005 2.087114 2.055551 0.5516497
#> [19,] 0.9065084 -0.17137633 2.0603949 2.259669 2.104748 2.452451 0.4093720
#> [20,] 0.9497075 -0.79635563 1.8592165 1.801996 1.862642 2.218350 0.5729377
#> [21,] 0.9416690 -0.39489538 1.6270096 2.188862 2.065925 1.951975 0.5343485
#> [22,] 0.7936251 -0.56921200 1.5677006 1.747128 1.913148 2.031674 0.5822931
#> [23,] 0.9949024 -0.53033917 1.4967439 1.823093 1.945096 1.962110 0.6256981
#> [24,] 0.9374582 -0.78128136 1.9561193 1.981729 1.942194 1.446557 0.5731984
#> [25,] 0.8102316 -0.41573745 1.6305938 2.272160 1.699281 1.710150 0.6446830
#> [26,] 0.8015282 -0.77973675 1.5861710 2.059627 1.725798 1.857511 0.6468463
#> [27,] 0.8632681 -0.15854210 1.6747503 2.087654 1.972693 2.018692 0.6214684
#> [28,] 0.8774739 -0.92125622 1.9415571 2.020442 1.755232 2.114356 0.5606285
#> [29,] 0.9982189 -0.36419881 1.9115883 2.053307 1.918857 1.880192 0.5164868
#> [30,] 0.8257987 -0.68037242 1.7859128 1.970513 1.991708 2.443009 0.5777355
#> [31,] 0.9370054 -0.70906507 2.0378245 2.069804 1.593629 2.201537 0.5871227
#> [32,] 0.8173915 -0.63283306 1.8121734 2.082570 1.776621 1.977370 0.5390020
#> [33,] 0.8876633 -0.87194317 1.6649749 2.105457 1.769212 1.664493 0.5584544
#> [34,] 0.9903892 -0.45414599 2.0071301 1.778924 2.151131 1.697312 0.4215273
#> [35,] 0.8968625 -0.60605727 1.7136295 1.955257 1.879813 1.924812 0.5684217
#> [36,] 0.9855938 -0.52790158 1.9692460 2.245370 1.705595 1.847121 0.6194299
#> [37,] 0.8307480 -0.62259797 1.9819094 1.890368 1.551287 1.733213 0.6471850
#> [38,] 0.8555451 -0.49430988 1.6511636 2.017434 2.234291 1.779078 0.5787997
#> [39,] 0.8348863 -0.45355954 1.6715798 2.007528 1.544036 1.881909 0.6567218
#> [40,] 0.7788404 -0.68614889 1.3752408 2.028497 1.565475 1.894037 0.7385256
#> [41,] 0.8829546 -0.74903408 1.4265420 1.910222 1.982988 1.957956 0.6309548
#> [42,] 0.9416412 -0.62273102 1.8732244 1.629123 1.834846 1.960098 0.5629723
#> [43,] 0.9016485 -0.44549311 1.7953858 2.202097 2.087385 1.427218 0.5194164
#> [44,] 0.9117249 -0.54188341 1.3482441 1.357965 1.892098 2.181001 0.6274892
#> [45,] 0.8798816 -0.13145228 1.5586832 2.039632 1.841876 2.102535 0.5557564
#> [46,] 0.9247249 -0.32031044 0.9099509 2.155004 2.176510 2.198915 0.6510300
#> [47,] 0.9167093 -0.18927359 1.2067775 2.244542 1.718660 2.016826 0.6755862
#> [48,] 0.9543394 -0.44457835 2.0948192 2.178138 2.052770 1.863414 0.4688680
#> [49,] 0.9104779 -0.53406778 2.1973335 1.974029 1.702744 2.110917 0.5830866
#> [50,] 0.8451026 -0.50799126 1.8997738 1.909238 1.726224 2.050496 0.5912897
#> [51,] 0.9535068  0.04804572 0.6361951 2.153520 1.483019 2.123886 0.7148478
#> [52,] 0.7949817 -0.66411701 1.3774441 1.625821 1.988544 1.479876 0.5937665
#> [53,] 0.9719629 -0.56555795 1.5230782 1.807771 1.845139 1.890548 0.6620727
#> [54,] 0.9382532 -0.79550227 1.7235027 2.000614 1.982200 2.245034 0.5669336
#> [55,] 0.9986435 -0.43607509 1.8277867 1.843831 2.058031 1.779489 0.4736578
#> [56,] 0.9519535 -0.29425317 1.9014164 3.235291 1.754654 1.991325 0.5492038
#> [57,] 0.8918441 -0.81805298 1.9469410 1.848479 1.883662 2.171592 0.5219383
#> [58,] 0.9734926 -0.94807452 1.7056852 1.955697 1.636016 1.873562 0.5441095
#> [59,] 0.8687403 -0.46029842 1.4099510 2.144719 1.632280 1.715390 0.6948125
#> [60,] 0.8793991 -0.35998100 1.8180920 1.855874 2.023458 1.884091 0.5831009
#> [61,] 0.8901810 -0.80268413 1.4928662 1.779478 1.745014 2.081836 0.5981379
#> [62,] 0.9199558 -0.59186099 1.5639323 1.994087 1.803117 1.617857 0.5803423
#> [63,] 0.9412222 -0.55083326 1.5289508 1.843986 1.878881 2.206690 0.5854939
#> [64,] 0.7348591 -0.61734473 1.8738494 1.664177 1.984776 2.095108 0.6362960
#> [65,] 0.9697437 -0.40447961 1.9916568 2.110722 1.920161 1.993846 0.6102171
#> [66,] 0.4870079 -0.54168402 1.5786331 2.025933 1.463050 2.230889 0.7016934
#> [67,] 0.9638148  0.19139653 1.7422411 1.827164 1.735077 1.737819 0.5677726
#> [68,] 0.9592531 -0.80780878 1.6024638 1.732346 1.832789 1.938439 0.6627059
#> [69,] 0.7521169 -0.52785517 1.2035703 2.156862 1.816555 2.124195 0.6212339
#> [70,] 0.9945453 -0.54765730 1.8718942 1.746613 2.051280 1.945771 0.5533877
#> [71,] 0.8887891 -0.78550872 1.4629308 1.730303 2.185133 2.466635 0.5469161
#> [72,] 0.9404248 -0.56574556 1.9893523 1.993181 1.706530 1.705223 0.5828746
#> [73,] 0.9027894 -0.91865042 1.5278775 1.944884 1.810992 2.101869 0.5916767
#> [74,] 0.9385192 -0.66727068 1.9550666 1.953947 1.714755 1.800445 0.6028970
#> [75,] 0.9206799 -0.59069183 1.6539965 1.745883 1.934230 1.834811 0.5980536
#> [76,] 0.9141714 -0.56977486 1.7325596 2.041531 2.186050 1.891604 0.5064705
#> [77,] 0.9816861 -0.65621246 1.7801932 2.168770 1.848633 2.177340 0.6314957
#> [78,] 0.8190865 -0.34751336 1.6963475 2.235921 1.645784 2.296439 0.5496913
#> [79,] 0.9537830 -0.55085732 1.6247045 2.020076 1.836402 1.883727 0.6421238
#> [80,] 0.8360316 -0.78600822 1.3985150 2.125389 1.869487 1.994004 0.5968116
#>            [,8]
#>  [1,] 0.5816618
#>  [2,] 0.5838870
#>  [3,] 0.4495008
#>  [4,] 0.6472545
#>  [5,] 0.5563873
#>  [6,] 0.6012055
#>  [7,] 0.5248445
#>  [8,] 0.4613800
#>  [9,] 0.5236572
#> [10,] 0.5226692
#> [11,] 0.5197359
#> [12,] 0.5097144
#> [13,] 0.4734306
#> [14,] 0.4445747
#> [15,] 0.5239244
#> [16,] 0.3713965
#> [17,] 0.5937001
#> [18,] 0.4897424
#> [19,] 0.3032630
#> [20,] 0.5989298
#> [21,] 0.5041797
#> [22,] 0.4876758
#> [23,] 0.4206398
#> [24,] 0.7023159
#> [25,] 0.5556742
#> [26,] 0.5662612
#> [27,] 0.3368171
#> [28,] 0.5249847
#> [29,] 0.4392106
#> [30,] 0.4708706
#> [31,] 0.5328686
#> [32,] 0.5953112
#> [33,] 0.4764618
#> [34,] 0.6499468
#> [35,] 0.5639565
#> [36,] 0.5315054
#> [37,] 0.5629949
#> [38,] 0.5668107
#> [39,] 0.5281095
#> [40,] 0.5403230
#> [41,] 0.5949768
#> [42,] 0.5281381
#> [43,] 0.6248761
#> [44,] 0.5764307
#> [45,] 0.4438566
#> [46,] 0.5164111
#> [47,] 0.4647616
#> [48,] 0.3197179
#> [49,] 0.3008883
#> [50,] 0.5162077
#> [51,] 0.5055907
#> [52,] 0.6822346
#> [53,] 0.5891680
#> [54,] 0.4907858
#> [55,] 0.5521251
#> [56,] 0.3741386
#> [57,] 0.5437598
#> [58,] 0.5468129
#> [59,] 0.4420038
#> [60,] 0.5135635
#> [61,] 0.5338423
#> [62,] 0.5460352
#> [63,] 0.4886076
#> [64,] 0.5519704
#> [65,] 0.4169913
#> [66,] 0.4069543
#> [67,] 0.5623494
#> [68,] 0.5894056
#> [69,] 0.4503016
#> [70,] 0.5499094
#> [71,] 0.5759844
#> [72,] 0.4789012
#> [73,] 0.5045149
#> [74,] 0.5706819
#> [75,] 0.5567500
#> [76,] 0.5263728
#> [77,] 0.5488465
#> [78,] 0.5093939
#> [79,] 0.5120314
#> [80,] 0.5022332
#> 
#> $member$storepop
#> list()
#> 
#> 
#> attr(,"class")
#> [1] "DEoptim"
#> 
#> $objective
#> [1] 292.3204
#> 
#> $residuals
#>               [,1]        [,2]
#>   [1,] -2.72507374 -0.86513842
#>   [2,]  3.18913858  0.79823464
#>   [3,] -0.92240390 -1.90226196
#>   [4,]  0.69735391 -1.22290487
#>   [5,]  1.35257125  1.02641480
#>   [6,] -0.78523682 -3.37315365
#>   [7,]  0.95903277  0.86310370
#>   [8,] -0.29792606 -0.43064422
#>   [9,] -0.56681139  0.67002727
#>  [10,]  0.26756332 -0.28979102
#>  [11,]  0.32319305 -1.47952135
#>  [12,]  0.49457283  0.40572563
#>  [13,] -0.41997426 -0.23088997
#>  [14,] -0.90871007 -1.13337940
#>  [15,] -1.37858593  0.12136806
#>  [16,]  0.40073415 -0.08752430
#>  [17,] -1.46429727  0.15329963
#>  [18,]  1.85303932 -0.68102583
#>  [19,]  0.43228112  0.51798859
#>  [20,] -0.45375684  0.82457444
#>  [21,]  0.88766833  1.09463295
#>  [22,] -0.98517540 -0.18589645
#>  [23,]  0.51479299  1.39508124
#>  [24,]  0.77338903 -2.19026476
#>  [25,] -0.28269273  0.61582069
#>  [26,] -0.99312183 -2.07501922
#>  [27,] -0.17094789 -0.50058725
#>  [28,] -0.85493883  0.09645089
#>  [29,] -0.52112488 -1.96657593
#>  [30,]  1.91235614  0.33790668
#>  [31,]  1.13735798  0.96712888
#>  [32,] -1.91664680 -1.52622557
#>  [33,] -1.19026861 -2.06638890
#>  [34,] -0.77975338  1.42125170
#>  [35,]  1.44731045  0.81202019
#>  [36,]  1.24929134 -0.91922202
#>  [37,] -0.56071678 -2.57904423
#>  [38,]  0.56005294  1.37374894
#>  [39,] -1.01132097 -1.96349573
#>  [40,]  1.06994349 -0.31196195
#>  [41,]  0.82692757  0.04338279
#>  [42,]  0.54823344  1.32610791
#>  [43,]  0.88459865 -0.37625173
#>  [44,] -0.06923060 -0.71344829
#>  [45,] -1.00171558 -0.73152313
#>  [46,] -0.84492297  0.86190386
#>  [47,]  1.04803703  2.30664434
#>  [48,] -1.39608565 -0.20395019
#>  [49,] -0.44951589 -0.95317028
#>  [50,] -0.17395264  1.18583114
#>  [51,] -0.39991797 -1.72350016
#>  [52,]  1.78225799  0.06235846
#>  [53,]  0.32653547 -0.84305563
#>  [54,] -1.20182320 -0.41992237
#>  [55,]  0.15728663  0.39089962
#>  [56,] -1.08088032  0.53655752
#>  [57,] -0.85843868 -1.16760347
#>  [58,] -1.41023017 -0.27563252
#>  [59,]  3.63660513  1.05017471
#>  [60,]  0.05836663 -1.51710028
#>  [61,] -0.97062302 -0.26376017
#>  [62,] -0.48183182 -0.88721042
#>  [63,] -0.55562467 -0.31150525
#>  [64,] -1.20828763 -2.30865430
#>  [65,]  0.98807542 -0.93942537
#>  [66,] -0.21063419 -2.13428545
#>  [67,] -0.18676928 -1.56585235
#>  [68,]  0.26972235  0.77318170
#>  [69,]  1.16637721  0.83595705
#>  [70,]  1.03361363 -0.12581321
#>  [71,]  1.81011330  1.07142234
#>  [72,]  0.10883369  0.98965331
#>  [73,]  0.95238299  0.39119866
#>  [74,] -0.02621775 -1.46132984
#>  [75,]  1.16325874  0.15968049
#>  [76,]  1.65975041  0.78140509
#>  [77,]  0.11656014 -0.02386481
#>  [78,]  1.87758728  0.99155625
#>  [79,] -1.83300624 -1.88597520
#>  [80,]  0.99090932  1.79732449
#>  [81,]  1.78548533  0.40737324
#>  [82,] -0.47775981 -0.52214263
#>  [83,]  0.64199901 -0.47119484
#>  [84,] -0.26510714  0.10397668
#>  [85,] -1.09171216  0.31025174
#>  [86,]  1.68111292  0.16744453
#>  [87,]  2.81170644  2.04311823
#>  [88,] -1.66310193 -0.21869761
#>  [89,] -0.08502568 -2.44185838
#>  [90,]  1.15383629  1.68879345
#>  [91,]  1.50458294  0.19119928
#>  [92,]  2.07585144  0.87467459
#>  [93,] -0.66700787  0.96174042
#>  [94,]  1.01712547 -0.38057113
#>  [95,]  2.09502331  0.76368293
#>  [96,]  1.26772040  2.41138040
#>  [97,]  1.39705380 -0.18572866
#>  [98,]  2.01812208  0.61499954
#>  [99,]  0.39608723  1.13525577
#> [100,]  0.35165963  2.81905515
#> 
#> $parameters
#>    alpha_1    alpha_2    beta_11    beta_21    beta_12    beta_22   gamma_11 
#>  0.8968625 -0.6060573  1.7136295  1.9552574  1.8798134  1.9248118  0.5684217 
#>   gamma_22   sigma_11   sigma_22 
#>  0.5639565  1.4172843  1.4713873 
#>